% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-element_textbox_colorsub.R
\name{element_grob.element_textbox_colorsub}
\alias{element_grob.element_textbox_colorsub}
\title{Custom ggplot2 element textbox highlight grob}
\usage{
element_grob.element_textbox_colorsub(element, label = "", ...)
}
\description{
Custom ggplot2 element textbox highlight grob
}
\seealso{
Other jam utility functions: 
\code{\link{element_textbox_colorsub}()},
\code{\link{get_se_assaydata}()},
\code{\link{handle_highlightPoints}()},
\code{\link{outer_legend}()},
\code{\link{points2polygonHull}()}
}
\concept{jam utility functions}
