% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-gg-blockarrow-geom.R
\name{geom_blockmargin}
\alias{geom_blockmargin}
\title{Block arrow or rectangle in the plot margins}
\usage{
geom_blockmargin(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  sides = "bl",
  direction = "u",
  length = grid::unit(0.05, "snpc"),
  arrow_type = "blockarrow",
  label = NULL,
  label_size = 18,
  label_color = NULL,
  font = 2,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this layer.
When using a \verb{geom_*()} function to construct a layer, the \code{stat}
argument can be used to override the default coupling between geoms and
stats. The \code{stat} argument accepts the following:
\itemize{
\item A \code{Stat} ggproto subclass, for example \code{StatCount}.
\item A string naming the stat. To give the stat as a string, strip the
function name of the \code{stat_} prefix. For example, to use \code{stat_count()},
give the stat as \code{"count"}.
\item For more information and other ways to specify the stat, see the
\link[ggplot2:layer_stats]{layer stat} documentation.
}}

\item{position}{A position adjustment to use on the data for this layer. This
can be used in various ways, including to prevent overplotting and
improving the display. The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example,
to use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation.
}}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}'s \code{params} argument. These
arguments broadly fall into one of 4 categories below. Notably, further
arguments to the \code{position} argument, or aesthetics that are required
can \emph{not} be passed through \code{...}. Unknown arguments that are not part
of the 4 categories below are ignored.
\itemize{
\item Static aesthetics that are not mapped to a scale, but are at a fixed
value and apply to the layer as a whole. For example, \code{colour = "red"}
or \code{linewidth = 3}. The geom's documentation has an \strong{Aesthetics}
section that lists the available options. The 'required' aesthetics
cannot be passed on to the \code{params}. Please note that while passing
unmapped aesthetics as vectors is technically possible, the order and
required length is not guaranteed to be parallel to the input data.
\item When constructing a layer using
a \verb{stat_*()} function, the \code{...} argument can be used to pass on
parameters to the \code{geom} part of the layer. An example of this is
\code{stat_density(geom = "area", outline.type = "both")}. The geom's
documentation lists which parameters it can accept.
\item Inversely, when constructing a layer using a
\verb{geom_*()} function, the \code{...} argument can be used to pass on parameters
to the \code{stat} part of the layer. An example of this is
\code{geom_area(stat = "density", adjust = 0.5)}. The stat's documentation
lists which parameters it can accept.
\item The \code{key_glyph} argument of \code{\link[ggplot2:layer]{layer()}} may also be passed on through
\code{...}. This can be one of the functions described as
\link[ggplot2:draw_key]{key glyphs}, to change the display of the layer in the legend.
}}

\item{sides}{\code{character} string to control which sides of the
plot the blocks appear. A string containing any letters in
\code{"trbl"} will set it to top, right, bottom and left respectively.}

\item{direction}{\code{character} vector with direction the arrow
should point. It recognizes letters in \code{'rlud'} as
right, left, up, and down, respectively.}

\item{length}{\code{grid::unit()} object that sets the width and
height of the rectangles in the x- and y-directions respectively.
Default is 5\% of the plot size, using \code{'snpc'} units so it
maintains consistent width for vertical and horizontal
block arrows. Change to \code{'npc'} to use fixed percentage
of the plot size.}

\item{arrow_type}{\code{character} indicating the type of drawing:
\itemize{
\item \code{"blockarrow"} for block arrow shapes, default.
\item \code{"rect"} is implemented, a simple rectangle.
}}

\item{label}{\code{character} text label to display on the rectangle.
Default \code{NULL}, or \code{NA} displays no label.}

\item{label_size}{\code{numeric} size for label text, default 18.}

\item{label_color}{\code{character} color for label text. Default NULL
uses a contrasting color with \code{jamba::setTextContrastColor()}.}

\item{font}{\code{integer} to define the font face, default 2 uses
bold font face, use 1 for normal.}

\item{na.rm}{\code{logical} not implemented.}

\item{show.legend}{\code{logical} or \code{NA}, not implemented.}

\item{inherit.aes}{\code{logical} not implemented.}
}
\value{
A \emph{Layer} ggproto object.
}
\description{
Draw rectangular blocks or block arrows in the plot margins,
suitable for annotating volcano plots with hit counts.
This is modeled after \code{ggh4x::geom_rectmargin()} but customized
for block arrow visualization.
}
\examples{
#\dontrun{
# Simple example with rectangles in margins
df <- data.frame(
  x = 1:10,
  y = rnorm(10),
  label = head(LETTERS, 10)
)
df
p <- ggplot2::ggplot(df, ggplot2::aes(x, y)) +
  ggplot2::geom_point() +
  geom_blockmargin(inherit.aes=FALSE,
    # ggplot2::aes(xmin = 2, xmax = 5, ymin = 0.5, ymax = 1.5),
    xmin = 2, xmax = 5, ymin = 0.5, ymax = 1.5,
    label = c("Label A", "Label B", rep("", 8)),
    fill = c("royalblue", "red", rep(NA, 8)),
    alpha = 0.3, sides = c("b", "l", rep("b", 8))
  ) +
  ggplot2::coord_cartesian(clip = "off")

print(p)
#}

}
