% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamma-se-data.R
\name{get_se_assaydata}
\alias{get_se_assaydata}
\title{Get SummarizedExperiment assay matrix data}
\usage{
get_se_assaydata(x, assay_name = NULL, verbose = FALSE, ...)
}
\arguments{
\item{x}{\code{SummarizedExperiment} object}

\item{assay_name}{\code{character} string that should match one entry
in \code{names(assays(x))}.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\description{
Get SummarizedExperiment assay matrix data
}
\details{
Simple function to generalize obtaining the assay data
from any object that inherits "SummarizedExperiment",
which includes for example "SingleCellExperiment",
"DESeqDataSet", "RangedSummarizedExperiment", "SpatialExperiment",
and others.

It also supports GeomxTools "NanoStringGeoMxSet" if the "GeomxTools"
R package is available.
}
\seealso{
Other jam utility functions:
\code{\link[=blockArrowMargin]{blockArrowMargin()}},
\code{\link[=find_colname]{find_colname()}},
\code{\link[=fold_to_log2fold]{fold_to_log2fold()}},
\code{\link[=get_se_colData]{get_se_colData()}},
\code{\link[=get_se_rowData]{get_se_rowData()}},
\code{\link[=gradient_rect]{gradient_rect()}},
\code{\link[=handle_highlightPoints]{handle_highlightPoints()}},
\code{\link[=log2fold_to_fold]{log2fold_to_fold()}},
\code{\link[=logAxis]{logAxis()}},
\code{\link[=outer_legend]{outer_legend()}},
\code{\link[=points2polygonHull]{points2polygonHull()}},
\code{\link[=update_function_params]{update_function_params()}},
\code{\link[=update_list_elements]{update_list_elements()}}
}
\concept{jam utility functions}
