% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamma-se-data.R
\name{get_se_colData}
\alias{get_se_colData}
\title{Get SummarizedExperiment column  data from colData}
\usage{
get_se_colData(x, use_values = NULL, verbose = FALSE, ...)
}
\arguments{
\item{x}{\code{SummarizedExperiment} object}

\item{use_values}{\code{character} values, default NULL. Optionally used
to validate input values in one of two ways:
\enumerate{
\item Vector of values typically named by \code{colnames(x)} that should
be applied to each column in \code{x} matching the order of the colnames.
\item Vector of values that matches \code{colnames(colData(x))}, which
should be converted to named vector of values, with names matching
\code{colnames(x)}.
}

Either way, when \code{use_values} is supplied, the output will be
a named vector, with names defined using \code{colnames(x)}.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{data.frame} or \code{DFrame} when \code{use_values} is not supplied,
or \code{character} vector named by \code{colnames(x)} when \code{use_values}
is supplied.
}
\description{
Get SummarizedExperiment column  data from colData
}
\examples{
\dontshow{if ({requireNamespace("SummarizedExperiment", quietly=TRUE)}) withAutoprint(\{ # examplesIf}
set.seed(123)
m <- matrix(rnorm(8100), ncol=9)
colnames(m) <- head(letters, 9)
rownames(m) <- as.character(1:900)
countse <- SummarizedExperiment::SummarizedExperiment(
   assays=list(counts=m))
SummarizedExperiment::colData(countse)$type <- rep(LETTERS[1:3], each=3)
SummarizedExperiment::colData(countse)$class <- rep(c("WT", "KO"), c(6, 3))

# use colname in colData
get_se_colData(countse, "type")

# use two colnames in colData
get_se_colData(countse, c("class", "type"))

# use three colnames in colData, one is missing
get_se_colData(countse, c("class", "type", "detail"))

# use named vector, which put it into proper order colnames(x)
use_values <- setNames(tail(LETTERS, 9), rev(head(letters, 9)))
use_values
get_se_colData(countse, use_values)

# use unnamed vector, which does not re-order values
unname(use_values)
get_se_colData(countse, unname(use_values))
\dontshow{\}) # examplesIf}
}
\seealso{
Other jam utility functions:
\code{\link[=blockArrowMargin]{blockArrowMargin()}},
\code{\link[=find_colname]{find_colname()}},
\code{\link[=fold_to_log2fold]{fold_to_log2fold()}},
\code{\link[=get_se_assaydata]{get_se_assaydata()}},
\code{\link[=get_se_rowData]{get_se_rowData()}},
\code{\link[=gradient_rect]{gradient_rect()}},
\code{\link[=handle_highlightPoints]{handle_highlightPoints()}},
\code{\link[=log2fold_to_fold]{log2fold_to_fold()}},
\code{\link[=logAxis]{logAxis()}},
\code{\link[=outer_legend]{outer_legend()}},
\code{\link[=points2polygonHull]{points2polygonHull()}},
\code{\link[=update_function_params]{update_function_params()}},
\code{\link[=update_list_elements]{update_list_elements()}}
}
\concept{jam utility functions}
