% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamma-ggjammaplot.R
\name{ggjammaplot}
\alias{ggjammaplot}
\title{MA-plots using ggplot2}
\usage{
ggjammaplot(
  jp2,
  x = NULL,
  useMedian = FALSE,
  controlSamples = NULL,
  centerGroups = NULL,
  groupedX = TRUE,
  grouped_mad = TRUE,
  outlierMAD = 5,
  mad_row_min = 4,
  displayMAD = FALSE,
  noise_floor = 1e-10,
  noise_floor_value = NA,
  naValue = NA,
  centerFunc = centerGeneData,
  whichSamples = NULL,
  useRank = FALSE,
  assay_name = 1,
  titleBoxColor = "lightgoldenrod1",
  outlierColor = "palegoldenrod",
  maintitle = NULL,
  subtitle = NULL,
  summary = "mean",
  difference = "difference",
  transFactor = 0.2,
  doPlot = TRUE,
  highlightPoints = NULL,
  highlightPch = 21,
  highlightCex = 1.5,
  highlightColor = NULL,
  doHighlightLegend = TRUE,
  nbin_factor = 1,
  ablineH = c(-2, 0, 2),
  base_size = 18,
  panel.grid.major.colour = "grey90",
  panel.grid.minor.colour = "grey95",
  return_type = c("ggplot", "data"),
  xlim = NULL,
  ylim = c(-6, 6),
  ncol = NULL,
  nrow = NULL,
  blankPlotPos = NULL,
  verbose = FALSE,
  ...
)
}
\description{
MA-plots using ggplot2
}
\details{
This method is under active development and may change as
features are implemented.

Currently the input to this function is the output of
\code{jammaplot()}.
}
\examples{
if (jamba::check_pkg_installed("SummarizedExperiment") &&
   jamba::check_pkg_installed("farrisdata")) {
   suppressPackageStartupMessages(require(SummarizedExperiment));

   GeneSE <- farrisdata::farrisGeneSE;

   titleBoxColor <- jamba::nameVector(
      farrisdata::colorSub[as.character(colData(GeneSE)$groupName)],
      colnames(GeneSE));

   jp2 <- jammaplot(GeneSE,
      outlierMAD=2,
      doPlot=FALSE,
      assay_name="raw_counts",
      filterFloor=1e-10,
      filterFloorReplacement=NA,
      centerGroups=colData(GeneSE)$Compartment,
      subtitleBoxColor=farrisdata::colorSub[as.character(colData(GeneSE)$Compartment)],
      useRank=FALSE);

   gg1 <- ggjammaplot(jp2,
      ncol=6,
      titleBoxColor=titleBoxColor);
   print(gg1);

   gg <- ggjammaplot(GeneSE, assay_name="raw_counts")

   gg <- ggjammaplot(GeneSE,
      assay_name="counts",
      useRank=TRUE,
      titleBoxColor=titleBoxColor)

   gg <- ggjammaplot(GeneSE,
      assay_name="counts",
      useRank=FALSE,
      titleBoxColor=titleBoxColor,
      ncol=6,
      displayMAD=TRUE)

   gg <- ggjammaplot(GeneSE,
      assay_name="counts",
      useRank=FALSE,
      titleBoxColor=titleBoxColor,
      ncol=6,
      whichSamples=colnames(GeneSE)[c(1:21, 23:24)],
      blankPlotPos=22,
      displayMAD=TRUE)

   ggdf <- ggjammaplot(GeneSE,
      assay_name="counts",
      whichSamples=c(1:3, 7:9),
      return_type="data",
      titleBoxColor=titleBoxColor)
   highlightPoints1 <- names(jamba::tcount(subset(ggdf, mean > 15 & difference < -1)$item, 2))
   highlightPoints2 <- subset(ggdf, name \%in\% "CA1CB492" &
      difference < -4.5)$item;
   highlightPoints <- list(
      divergent=highlightPoints1,
      low_CA1CB492=highlightPoints2);

   ggdf_h <- ggjammaplot(GeneSE,
      assay_name="counts",
      highlightPoints=highlightPoints,
      whichSamples=c(1:3, 7:9),
      return_type="data",
      titleBoxColor=titleBoxColor)

}

}
