% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamma-calc.R
\name{jammacalc}
\alias{jammacalc}
\title{Calculate MA-plot data}
\usage{
jammacalc(
  x,
  na.rm = TRUE,
  controlSamples = NULL,
  centerGroups = NULL,
  controlFloor = NA,
  naControlAction = c("row", "floor", "min", "na"),
  naControlFloor = 0,
  groupedX = TRUE,
  useMedian = TRUE,
  useMean = NULL,
  whichSamples = NULL,
  noise_floor = -Inf,
  noise_floor_value = noise_floor,
  naValue = NA,
  mad_row_min = 0,
  grouped_mad = TRUE,
  centerFunc = centerGeneData,
  useRank = FALSE,
  returnType = c("ma_list", "tidy"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{\code{numeric} matrix typically containing log-normal measurements,
with measurement rows, and sample columns.}

\item{na.rm}{\code{logical} indicating whether to ignore NA values
during \code{numeric} summary functions.}

\item{controlSamples}{\code{character} vector containing values in
\code{colnames(x)} to define control samples used during centering.
These values are passed to \code{centerGeneData()}.}

\item{centerGroups}{\code{character} vector with length equal to \code{ncol(x)}
which defines the group for each column in \code{x}. Data will
be centered within each group.}

\item{groupedX}{\code{logical} indicating how to calculate the x-axis
value when \code{centerGroups} contains multiple groups. When
groupedX=TRUE, the mean of each group median is used, which
has the effect of representing each group equally. When
groupedX=FALSE, the median across all columns is used, which
can have the effect of preferring sample groups with a larger
number of columns.}

\item{useMedian}{\code{logical} indicating whether to use the median
values when calculating the x-axis and during data centering.
The median naturally reduces the effect of outlier points on
the resulting MA-plots., when compared to using the mean.
When useMedian=FALSE, the mean value is used.}

\item{useMean}{(deprecated) \code{logical} indicating whether to use the
mean instead of the median value. This argument is being removed
in order to improve consistency with other Jam package functions.}

\item{whichSamples}{\code{character} vector containing \code{colnames(x)}, or
integer vector referencing column numbers in \code{x}. This argument
specifies which columns to return, but does not change the columns
used to define the group centering values. For example, the
group medians are calculated using all the data, but only the
samples in \code{whichSamples} are centered to produce MA-plot data.}

\item{noise_floor}{\code{numeric} value indicating the minimum numeric value
allowed in the input matrix \code{x}. When \code{NULL} or \code{-Inf} no noise
floor is applied. It is common to set \code{noise_floor=0} to limit
MA-plot data to use values zero and above.}

\item{noise_floor_value}{single \code{numeric} value used to replace \code{numeric}
values at or below \code{noise_floor} when \code{noise_floor} is not NULL.
By default,
\code{noise_floor_value=noise_floor} which means values at or below
the noise floor are set to the floor. Another useful option is
\code{noise_floor_value=NA} which has the effect of removing the point
from the MA-plot altogether. This option is recommended for sparse
data matrices where the presence of values at or below zero are
indicative of missing data (zero-inflated data) and does not
automatically reflect an actual value of zero.}

\item{naValue}{single \code{numeric} value used to replace any \code{NA} values in
the input matrix \code{x}. This argument can be useful to replace
\code{NA} values with something like zero.}

\item{mad_row_min}{\code{numeric} value defining the minimum group
value, corresponding to the x-axis position on the MA-plot,
required for a row to be included in the MAD calculation.
This threshold is useful to filter outlier data below a noise
threshold, so that the MAD calculation will include only the
data above that value. For example, with count data, it is
useful to filter out counts below roughly 8, where Poisson
noise is a more dominant component than real count data.
Remember that count data should already be log2-transformed,
so the threshold should also be identically transformed,
for example using \code{log2(1 + 8)} to set a minimum count
threshold of at least 8.}

\item{grouped_mad}{\code{logical} indicating whether the MAD value
should be calculated per group when \code{centerGroups} is
supplied, from which the MAD factor values are derived.
When \code{TRUE} it has the effect of highlighting outliers
within each group using the variability in that group.
When \code{FALSE} the overall MAD is calculated, and a
particularly high variability group may have all its
group members labeled with a high MAD factor.}

\item{centerFunc}{\code{function} used for centering data, by default
one of the functions \code{centerGeneData()} or \code{centerGeneData_v1()}.
This argument will be removed in the near future and is mainly
intended to allow testing the two centering functions.
The following arguments are passed to this function:
\itemize{
\item x: the input \code{numeric} data matrix
\item na.rm: \code{logical} whether to ignore NA value. Always use \code{na.rm=TRUE}.
\item controlSamples: \code{character} optional subset of \code{colnames(x)} to
use as reference controls during centering
\item centerGroups: \code{character} vector of groups for \code{colnames(x)}
\item controlFloor: \code{numeric} optional minimum allowed value for control
summary prior to centering
\item naControlAction: \code{character} string for how to handle entirely NA
control groups during centering
\item naControlFloor: \code{numeric} used when \code{naControlAction="floor"} and
all control values are \code{NA}. One \code{numeric} value is inserted into
the control group.
\item useMedian: \code{logical} whether to use median (TRUE) or mean (FALSE)
\item returnGroups: \code{logical} whether to return summary of group assignment
in attribute \code{"center_df"}
\item returnGroupedValues: \code{logical} whether to return group summary values
in attribute \code{"x_group"}
\item ...: other arguments are passed along via \code{...}.
}}

\item{returnType}{\code{character} string indicating the format of data
to return: \code{"ma_list"} is a list of MA-plot two-column
numeric matrices with colnames \code{c("x","y")}; "tidy"
returns a tall \code{data.frame} suitable for use in ggplot2.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\description{
Calculate MA-plot data
}
\details{
This function takes a numeric matrix as input, and calculates
data sufficient to produce MA-plots. The default output is a
list of two-column numeric matrices with \code{"x"} and \code{"y"} coordinates,
representing the group median and difference from median,
respectively.

The mean value can be used by setting \code{useMedian=FALSE}.

Samples can be grouped using the argument \code{centerGroups}.
In this case the y-axis value will be "difference from
group median."

Control samples can be specified for centering using the
argument \code{controlSamples}. In this case, the y-axis value will
be "difference from control median".

The sample grouping, and control samples can be combined,
in which case the y-axis values will be "difference from
the control median within the centering group."
}
\seealso{
Other jam matrix functions:
\code{\link[=centerGeneData]{centerGeneData()}},
\code{\link[=jammanorm]{jammanorm()}},
\code{\link[=matrix_to_column_rank]{matrix_to_column_rank()}}
}
\concept{jam matrix functions}
