% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-log2fold-to-fold.R
\name{log2fold_to_fold}
\alias{log2fold_to_fold}
\title{Convert log2 fold change to signed fold change}
\usage{
log2fold_to_fold(x, ...)
}
\arguments{
\item{x}{\code{numeric} vector}
}
\value{
\code{numeric} vector representing signed fold change values.
}
\description{
Convert log2 fold change to signed fold change
}
\details{
This function takes log2 fold change values as input, and returns
normal space fold change values that retain the positive and negative
sign, and the magnitude.

For example:
\itemize{
\item \verb{log2 fold change = 2} becomes \verb{fold change = 4}.
\item \verb{log2 fold change = -2} becomes \verb{fold change = -4}.
}

This function therefore differs from similar functions that convert
log2 fold change into a ratio. Instead, \code{log2fold_to_fold()}
specifically retains the magnitude of negative changes.
}
\examples{
x <- c(-3, -2, -1, 0, 1, 2, 3);
fc <- log2fold_to_fold(x);
fc;

fold_to_log2fold(fc);

}
\seealso{
Other jam utility functions:
\code{\link[=blockArrowMargin]{blockArrowMargin()}},
\code{\link[=find_colname]{find_colname()}},
\code{\link[=fold_to_log2fold]{fold_to_log2fold()}},
\code{\link[=get_se_assaydata]{get_se_assaydata()}},
\code{\link[=get_se_colData]{get_se_colData()}},
\code{\link[=get_se_rowData]{get_se_rowData()}},
\code{\link[=gradient_rect]{gradient_rect()}},
\code{\link[=handle_highlightPoints]{handle_highlightPoints()}},
\code{\link[=logAxis]{logAxis()}},
\code{\link[=outer_legend]{outer_legend()}},
\code{\link[=points2polygonHull]{points2polygonHull()}},
\code{\link[=update_function_params]{update_function_params()}},
\code{\link[=update_list_elements]{update_list_elements()}}
}
\concept{jam utility functions}
