% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamma-numeric-to-column-rank.R
\name{matrix_to_column_rank}
\alias{matrix_to_column_rank}
\title{Convert numeric matrix to column-rank data}
\usage{
matrix_to_column_rank(x, keepNA = TRUE, usePercentile = FALSE, ...)
}
\arguments{
\item{x}{\code{numeric} matrix data}

\item{...}{additional arguments are ignored.}
}
\value{
\code{numeric} matrix with same dimensions as input data.
}
\description{
Convert numeric matrix to column-rank data
}
\details{
This function converts a numeric matrix to rank-based data,
where the higher rank value refers to higher numeric value in
the input data. It therefore behaves similar to continuous
numeric data:
\itemize{
\item higher values indicate higher signal
\item difference in values indicates difference in rank
}
}
\seealso{
Other jam matrix functions:
\code{\link[=centerGeneData]{centerGeneData()}},
\code{\link[=jammacalc]{jammacalc()}},
\code{\link[=jammanorm]{jammanorm()}}
}
\concept{jam matrix functions}
