% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-misc.R
\name{outer_legend}
\alias{outer_legend}
\title{Add legend in the outer margin}
\usage{
outer_legend(
  x = "bottom",
  y = NULL,
  doPar = TRUE,
  legend,
  bg = "white",
  box.col = "grey70",
  col,
  pt.bg = NULL,
  pch = 21,
  ncol = min(c(5, length(legend))),
  cex = 1.2,
  pt.cex = 1.4,
  ...
)
}
\arguments{
\item{x, y}{position of the legend, using any method
recognized by \code{grDevices::xy.coords()}.}

\item{doPar}{logical indicating whether to apply
\code{graphics::par()} in order to place the legend in
the outer margin. If \code{doPar=FALSE} and if no other
method has been done to update the \code{graphics::par()}
settings, the legend will be placed with the same
mechanism used by \code{graphics::legend()}.}

\item{legend}{character vector, or expression, used to
display each legend label.}

\item{bg}{color used as the background of the legend box.}

\item{box.col}{color used as the frame border color of the
legend box.}

\item{col}{\code{character} vector of colors used to color each
entry in \code{legend}.}

\item{pt.bg}{vector of colors used as the background color
when \code{pch} is any value from 21 to 25. When \code{pt.bg=NULL}
this value and \code{col} are adjusted so the outline is
darker, via \code{jamba::makeColorDarker()}, with the fill
color \code{pt.bg} using the original \code{col} value.}

\item{pch}{integer or character vector of point shapes,
as described in \code{graphics::points()}.}

\item{ncol}{integer number of columns to use for legend
entries. By default \code{ncol} is the smaller of
\code{length(legend)} and \code{5}, which is intended for a legend
placed at the bottom of the visualization.}

\item{cex}{character expansion multiplier, used to size
the legend text overall.}

\item{pt.cex}{point expansion factor, used to size points
when \code{pch} is not \code{NULL}.}

\item{...}{additional arguments are passed to
\code{graphics::legend()}.}
}
\description{
Add legend in the outer margin
}
\details{
This function is intended to place a legend outside the figure
in the outer margin area. The motivating use case is to
display a legend below multi-panel base R graphics plots,
for example when using \code{par("mfrow"=c(2,2))}.
}
\examples{
opar <- par(no.readonly=TRUE);
on.exit(par(opar));

par("mfrow"=c(2,2));
for (i in 1:4) {
   jamba::nullPlot(plotAreaTitle=paste("Panel", i));
}
require(jamba);
outer_legend("bottom",
   legend=c("one", "two", "three"),
   col=colorjam::rainbowJam(3));

}
\seealso{
Other jam utility functions:
\code{\link[=blockArrowMargin]{blockArrowMargin()}},
\code{\link[=find_colname]{find_colname()}},
\code{\link[=fold_to_log2fold]{fold_to_log2fold()}},
\code{\link[=get_se_assaydata]{get_se_assaydata()}},
\code{\link[=get_se_colData]{get_se_colData()}},
\code{\link[=get_se_rowData]{get_se_rowData()}},
\code{\link[=gradient_rect]{gradient_rect()}},
\code{\link[=handle_highlightPoints]{handle_highlightPoints()}},
\code{\link[=log2fold_to_fold]{log2fold_to_fold()}},
\code{\link[=logAxis]{logAxis()}},
\code{\link[=points2polygonHull]{points2polygonHull()}},
\code{\link[=update_function_params]{update_function_params()}},
\code{\link[=update_list_elements]{update_list_elements()}}
}
\concept{jam utility functions}
