% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamma.r
\name{points2polygonHull}
\alias{points2polygonHull}
\title{define a polygon hull around points}
\usage{
points2polygonHull(x, returnClass = c("matrix", "SpatialPolygons"), ...)
}
\arguments{
\item{x}{two-column numeric matrix of data points, or an object of
class \code{sp::SpatialPoints}.}

\item{returnClass}{character string indicating the class to return,
either \code{"matrix"} to return a two-column numeric matrix, or
\code{"SpatialPolygons"} to return \code{sp::SpatialPolygons}.}

\item{verbose}{logical whether to print verbose output.}
}
\value{
A two-column numeric matrix of polygon coordinates when
\code{returnClass="matrix"} by default, or when
\code{returnClass="SpatialPolygons"} it returns an object
of class \code{sp::SpatialPolygons}.
}
\description{
Define a polygon hull to encompass a set of data points
}
\details{
A polygon hull can be useful as a visual indicator of the location
of a set of points. This function creates a polygon around all points,
and not just the most dense region of points, using
\code{grDevices::chull()}.

The output is slightly modified to duplicate the first coordinate,
in order to "close" the resulting polygon, which is consistent
with the output of \code{rgeos::gConvexHull()}.
}
\examples{
set.seed(123);
xy <- cbind(x=1:10,
   y=sample(1:10, 10));
xyhull <- points2polygonHull(xy);
plot(xy, pch=20, cex=1, col="purple4",
   main="Polygon hull around points");
polygon(xyhull, border="purple2", col="transparent");

}
\seealso{
Other jam utility functions:
\code{\link[=blockArrowMargin]{blockArrowMargin()}},
\code{\link[=find_colname]{find_colname()}},
\code{\link[=fold_to_log2fold]{fold_to_log2fold()}},
\code{\link[=get_se_assaydata]{get_se_assaydata()}},
\code{\link[=get_se_colData]{get_se_colData()}},
\code{\link[=get_se_rowData]{get_se_rowData()}},
\code{\link[=gradient_rect]{gradient_rect()}},
\code{\link[=handle_highlightPoints]{handle_highlightPoints()}},
\code{\link[=log2fold_to_fold]{log2fold_to_fold()}},
\code{\link[=logAxis]{logAxis()}},
\code{\link[=outer_legend]{outer_legend()}},
\code{\link[=update_function_params]{update_function_params()}},
\code{\link[=update_list_elements]{update_list_elements()}}
}
\concept{jam utility functions}
