% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamma-volcano.R
\name{resizeSpatialPolygons}
\alias{resizeSpatialPolygons}
\title{Resize sp SpatialPolygons object}
\usage{
resizeSpatialPolygons(obj, size = c(1, 1), offsetXY = c(0, 0), ...)
}
\description{
Resize sp SpatialPolygons object
}
