% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamma-volcano.R
\name{trimAngle}
\alias{trimAngle}
\title{Trim angle to a fixed range}
\usage{
trimAngle(x, minAngle = 0, maxAngle = NULL, unit = c("degrees",
  "radians"), ...)
}
\description{
Trim angle to a fixed range
}
