% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-update-fn-params.R
\name{update_list_elements}
\alias{update_list_elements}
\title{Update a subset of list elements (deprecated)}
\usage{
update_list_elements(
  source_list,
  update_list,
  list_layer_num = 1,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{source_list}{\code{list} with input values}

\item{update_list}{\code{list} with values to populate into \code{source_list}.}

\item{list_layer_num}{\code{integer} used for internal use by this function,
which defines the list depth when traversing a nested list.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\description{
Update a subset of list elements (deprecated)
}
\details{
Deprecated.

This function is intended to help update a nested \code{source_list},
a subset of whose values should be replaced with entries
in \code{update_list}, leaving any original entries in \code{source_list}
which were not defined in \code{update_list}.

This function may be useful when manipulating lattice or ggplot2
graphical parameters, which are often stored in a nested
list structure.

Note: This function is deprecated, because there exists a function
\code{utils::modifyList()} written by Lattice co-author
Dr. Deepayan Sarkar, that serves the same purpose as this function.
No doubt the comment above is accurate, that this function is
widely used in the Lattice R package to update graphical
parameters.
}
\examples{
# function starts with y formals, adds or updates new values
y <- list(a=1, b=4, e=NULL)
update_list_elements(y, list(b=5:6, d=1:5));

# also see the preferred function utils::modifyList below
utils::modifyList(y, list(b=5:6, d=1:5));

}
\seealso{
Other jam utility functions:
\code{\link[=blockArrowMargin]{blockArrowMargin()}},
\code{\link[=find_colname]{find_colname()}},
\code{\link[=fold_to_log2fold]{fold_to_log2fold()}},
\code{\link[=get_se_assaydata]{get_se_assaydata()}},
\code{\link[=get_se_colData]{get_se_colData()}},
\code{\link[=get_se_rowData]{get_se_rowData()}},
\code{\link[=gradient_rect]{gradient_rect()}},
\code{\link[=handle_highlightPoints]{handle_highlightPoints()}},
\code{\link[=log2fold_to_fold]{log2fold_to_fold()}},
\code{\link[=logAxis]{logAxis()}},
\code{\link[=outer_legend]{outer_legend()}},
\code{\link[=points2polygonHull]{points2polygonHull()}},
\code{\link[=update_function_params]{update_function_params()}}
}
\concept{jam utility functions}
