% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-volcano-plot.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Volcano plot}
\usage{
volcano_plot(
  x,
  n = NULL,
  lfc_colname = c("logfc", "log2fold", "log2fc", "lfc", "l2fc", "logratio", "log2ratio"),
  fold_colname = c("fold", "fc", "ratio"),
  fold_style = c("fold", "log2fold", "log2", "lfc"),
  fold_cutoff = 1.5,
  fold_max_range = 16,
  fold_min_range = 4,
  sig_colname = c("adj.P.Val", "padj", "adj.pval", "adjp", "P.Value"),
  sig_cutoff = 0.05,
  sig_max_range = 1e-10,
  sig_min_range = 1e-04,
  expr_colname = c("mgm", "groupmean", "mean", "AveExpr", "fkpm", "rpkm", "tpm", "cpm"),
  expr_cutoff = NULL,
  gene_colname = c("gene", "symbol", "protein", "probe", "assay", "rownames"),
  label_colname = c("gene", "symbol", "protein", "probe", "assay", "rownames"),
  main = "Volcano Plot",
  submain = NULL,
  blockarrow = TRUE,
  blockarrow_colors = c(hit = "#E67739FF", up = "#990000FF", down = "#000099FF"),
  blockarrow_font = 1,
  blockarrow_cex = c(1.2, 1.2),
  blockarrow_label_cex = 1,
  blockarrow_shadowtext = TRUE,
  symmetric_axes = TRUE,
  do_cutoff_caption = TRUE,
  caption_cex = 0.8,
  include_axis_prefix = FALSE,
  n_x_labels = 12,
  n_y_labels = 7,
  xlim = NULL,
  ylim = NULL,
  pt_cex = 0.9,
  pt_pch = 21,
  hit_type = "hits",
  color_set = c(base = "#77777777", up = "#99000088", down = "#00009988", hi =
    "#FFDD55FF", hi_up = "#FFDD55FF", hi_down = "#FFDD55FF"),
  border_set = NULL,
  point_colors = NULL,
  border_colors = NULL,
  abline_color = "#000000AA",
  smooth = TRUE,
  smooth_func = jamba::plotSmoothScatter,
  smooth_ramp = colorRampPalette(c("white", "lightblue", "lightskyblue3", "royalblue",
    "darkblue", "orange", "darkorange1", "orangered2")),
  tophist = FALSE,
  tophist_cutoffs = c("pvalue", "foldchange"),
  tophist_breaks = 100,
  tophist_color = "#000099FF",
  tophist_fraction = 1/3,
  tophist_by = 0.2,
  hi_points = NULL,
  hi_colors = NULL,
  hi_hits = FALSE,
  hi_cex = 1,
  hi_do_label = FALSE,
  max_labels = 50,
  do_both = FALSE,
  label_hits = FALSE,
  add_plot = FALSE,
  do_plot = TRUE,
  xlab = NULL,
  ylab = NULL,
  cex.axis = 1.2,
  mar_min = c(6, 5, 6, 5),
  transFactor = 0.24,
  transformation = function(x) {
     x^transFactor
 },
  nbin = 256,
  bw_factor = 1,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{\code{data.frame} that contains statistical results with at
least a P-value, and fold change or log2 fold change. It is
useful to contain a column with mean expression, and a column
with a relevant label.}

\item{n}{\code{integer} indicating the number of subset points to plot
for testing purposes.}

\item{lfc_colname}{\code{character} string or vector used to match
\code{colnames(x)} whose values should be log2 fold changes.
A direct match to \code{colnames(x)} is performed
first, then if no column is found, the values are used as
regular expression patterns in order until the first
matching colname is found. Note that \code{lfc_colname} is used
in preference to \code{fold_colname}.
The colname used will appear as the x-axis label.}

\item{fold_colname}{\code{character} string or vector used to match
\code{colnames(x)} whose values should be fold changes. Note that
if \code{lfc_colname} successfully finds a value, the \code{fold_colname}
is not used.
The colname if used will appear as the x-axis label.}

\item{fold_style}{\code{character} style of x-axis label, default 'fold':
\itemize{
\item 'fold': shows normal space fold change, on log2 visual scale.
For example 1, 2, 4, 8, 16.
\item 'log2fold', 'log2', 'log', or 'lfc': shows log2 fold change.
For example 0, 1, 2, 3, 4.
}}

\item{fold_cutoff}{\code{numeric} threshold for values in \code{lfc_colname}
or \code{fold_cutoff}, where normal fold change values at or above
\code{fold_cutoff} can be considered statistically significant.
Note that when \code{lfc_colname} is being used, its values are
converted to normal fold change before applying this filter.}

\item{fold_max_range}{\code{numeric} indicating the maximum range to display
on the x-axis fold change. This argument prevents extremely
large fold changes from compressing the useful visible range
of the figure.}

\item{fold_min_range}{\code{numeric} indicating the minimum range to display
on the x-axis fold change. This argument is useful
when fold changes are low and the x-axis range would otherwise
be too small to be very useful.}

\item{sig_colname}{\code{character} string or vector used to match
\code{colnames(x)} whose values should contain P-values of significance.
The P-values can be unadjusted (raw) P-values, or adjusted
P-values. The P-values are expected not to be \code{-log10()}
transformed.
The colname used will appear as the y-axis label.}

\item{sig_cutoff}{\code{numeric} threshold for values in \code{sig_colname},
where values at or below \code{sig_cutoff} can be considered
statistically significant.}

\item{sig_max_range}{\code{numeric} indicating the maximum range to display
on the y-axis significance. This argument prevents extremely
small P-values from compressing the useful visible range
of the figure.}

\item{sig_min_range}{\code{numeric} indicating the minimum range to display
on the y-axis significance. This argument is useful
when P-values are not very significant, and you want to make
sure the y-axis range shows a minimum amount of the significant
region to be visually interpretable in that context.}

\item{expr_colname}{\code{character} string or vector used to match
\code{colnames(x)} whose values should contain expression mean values.
This column is only used when \code{expr_cutoff} is defined and
is applied to the filter criteria for statistical hits.}

\item{expr_cutoff}{\code{numeric} threshold for values in \code{expr_colname}
when \code{expr_colname} is defined, where values in \code{expr_colname}
at or above \code{expr_cutoff} can be considered statistically
significant. This threshold is useful to filter out potential
statistical hits whose signal is below a noise signal threshold.}

\item{label_colname}{\code{character} string or vector used to match
\code{colnames(x)} whose values should contain a useful label,
for example gene symbol or assay identifier.}

\item{main}{\code{character} string used as the main title of the figure.}

\item{submain}{\code{character} string used as a sub-title of the figure.}

\item{blockarrow}{\code{logical} indicating whether block arrows should
be displayed and used to indicate the number of statistical hits.}

\item{blockarrow_colors, blockarrow_font, blockarrow_label, blockarrow_shadowtext}{arguments used when \code{blockarrow=TRUE}.}

\item{symmetric_axes}{\code{logical} indicating whether the x-axis
log fold change range should be symmetric above and below zero.}

\item{do_cutoff_caption}{\code{logical} indicating whether to display
text caption with the statistical cutoff values used, and the
total number of points displayed.}

\item{caption_cex}{\code{numeric} caption font size adjustment.}

\item{include_axis_prefix}{\code{logical} indicating whether to include
a prefix for the x-axis and y-axis labels: x-axis \code{"Change"};
y-axis \code{"Significance"}.}

\item{n_x_labels, n_y_labels}{\code{integer} used by \code{pretty()} to determine
the approximate number of x-axis and y-axis labels to display,
respectively.}

\item{xlim, ylim}{\code{numeric} used to define specific \code{xlim} and \code{ylim}
axis ranges. When \code{NULL} the ranges are defined automatically,
using \code{fold_min_range},\code{fold_max_range} for the x-axis, and
\code{sig_min_range},\code{sig_max_range} for the y-axis.}

\item{pt_cex, pt_pch}{\code{numeric} used to define point size and shape,
used only when individual points are displayed.}

\item{hit_type}{\code{character} string used to label points that meet
the statistical cutoffs as \code{"hits"}, but where it may be useful
to indicate the type of entry being tested. For example:
\itemize{
\item \code{hit_type="genes"} indicates that each row represents a gene;
\item \code{hit_type="probes"} indicates each row represents a probe;
\item \code{hit_type="transcripts"} indicates each row represents a transcript.
}}

\item{color_set}{\code{character} vector of R colors, used only when individual
points are display. The names override default values, and may include:
\itemize{
\item \code{"base"} - the base color of all points on the plot
\item \code{"up"} - the color for up-regulated points that meet all
statistical cutoffs to be a "hit".
\item \code{"down"} - the color for down-regulated points that meet all cutoffs
\item \code{"hi"} - base color for highlighted points, used when \code{hi_points}
is defined.
\item \code{"hi_up"} - color for highlighted up-regulated points.
\item \code{"hi_down"} - color for highlighted down-regulated points.
}}

\item{border_set}{\code{NULL} or \code{character} vector of R colors, used to
define point border colors such as \code{pch=21} which is a filled circle
with border. When \code{border_set=NULL} then it is defined by
\code{jamba::makeColorDarker(color_set)}.}

\item{point_colors, border_colors}{optional \code{character} vector of R colors
recycled to length \code{nrow(x)}, used to specify the exact color of each
point in \code{x}. This argument is useful to colorize certain specific
points that may otherwise not meet statistical criteria.}

\item{abline_color}{\code{character} string with R color used to color
the abline that indicates the x-axis \code{fold_cutoff} value, and
y-axis \code{sig_cutoff} value.}

\item{smooth}{\code{logical} indicating whether points should be drawn
as a smooth scatter plot, using \code{jamba::plotSmoothScatter()}.
When \code{smooth=FALSE} individual points are drawn, using
\code{point_colors}, or when \code{point_colors} is not defined the
default is to use \code{color_set} to colorize points based upon
statistical cutoffs.}

\item{smooth_func}{\code{function} used to plot points when \code{smooth=TRUE},
by default \code{jamba::plotSmoothScatter()} which has some benefits
over default \code{graphics::smoothScatter()}.}

\item{smooth_ramp}{\code{character} vector of R colors which defines
the color gradient to use when \code{smooth=TRUE}.}

\item{tophist}{\code{logical} indicating whether to display a histogram
at the top of the volcano plot figure.}

\item{tophist_cutoffs, tophist_breaks, tophist_fraction, tophist_by}{arguments used when \code{tophist=TRUE}.}

\item{hi_points}{\code{character} vector indicating points to highlight
in the volcano plot, where values should match \code{rownames(x)}.
This argument is useful to highlight a specific subset of points of
interest on the figure. Note that \code{hi_points} are always
rendered as individual points even when \code{smooth=TRUE}.}

\item{hi_hits}{\code{logical} indicating whether rows that meet all
statistical cutoffs and are considered "hits" should also be
treated as \code{hi_points} for the purpose of rendering individual
points.}

\item{hi_cex}{\code{numeric} size adjustment for highlight points,
relative to the size of other points in the figure.}

\item{hi_do_label}{\code{logical} with length equivalent to \code{hi_points},
and recycled to match length or lengths of hi_points.}

\item{max_labels}{\code{integer} maximum labels to display when
\code{hi_do_label} is enabled. When the total potential labels is
more than \code{max_labels}, a subset of points will be shown based
upon point density, similar to how \code{smoothScatter()} argument
'nrpoints' chooses extreme points at the edges of the point
density. Note the point density uses the full range of values,
not the xlim and ylim ranges, in order to include the most
relevant extremes.}

\item{do_both}{\code{logical} indicating whether to draw both a smooth
scatter and individual points on the same figure.}

\item{label_hits}{\code{logical} indicating whether to add a text label
for points that are statistical hits.}

\item{add_plot}{\code{logical} indicating whether the plot should be
added to an existing plot, or when \code{add_plot=FALSE} a new
plot is created. This argument is useful to re-run the
same volcano plot with alternate parameters, for example
to display different subsets of highlighted points.}

\item{xlab, ylab}{\code{character} strings used to specify the exact
x-axis label and y-axis label. When either value is \code{NULL}
the default is to use the relevant colname: x-axis uses
either \code{lfc_colname} or \code{fold_colname}; y-axis uses \code{sig_colname}.}

\item{cex.axis}{\code{numeric} adjustment for axis label font sizes.}

\item{mar_min}{\code{numeric} vector to set minimum plot margins for base
R output. Default is c(6, 5, 6, 5) in units of character lines,
and in order: bottom, left, top, right.
Note when \code{blockarrow=FALSE} the top and right margins are subtracted
by 2, imposing a minimum \code{0.1}.}

\item{transformation}{\code{function} passed to \code{smooth_func} used to
adjust the visual contrast of the resulting density plot.}

\item{nbin}{\code{numeric} value passed to \code{smooth_func} and used
by \code{jamba::plotSmoothScatter()} to adjust the number of
bins used to display the density of points, where a higher
value shows more detail, and a lower value shows less detail.}

\item{bw_factor}{\code{numeric} scalar to adjust the bandwidth relative
detail, with larger numbers displaying more granular detail in
the point density shown. It is passed to \code{jamba::plotSmoothScatter()}
in the form of adjusting the \code{bwpi} bandwidth-per-inch.}

\item{verbose}{\code{logical} indicating whether to print verbose output.
Note that \code{verbose=2} will enable much more verbose output.}

\item{...}{additional arguments are ignored.}

\item{numeric}{vector used to ensure that each margin size is
at least a minimum value, applied to \code{par("mar")} via
the function \code{pmax()}.}
}
\description{
Draw a volcano plot using reasonable default arguments.
}
\details{
Draw a volcano plot using a reasonably robust set of default
arguments, and with a large number of customization options.
The default plot uses smooth scatter plot for much improved
display of point density.

This function produces a volcano plot, which consists of
change on the x-axis, and significance on the y-axis.

In addition to displaying the volcano plot, this function
also displays statistical thresholds, and marks entries
as "hits" by up to three conceptual filters:
\itemize{
\item "change" - fold change \code{fold_cutoff}
\item "significant" - statistical P-value \code{sig_cutoff}
\item "detected" - signal \code{expr_cutoff}
}

If any cutoff is not defined, that filter is ignored.

Change is usually represented using log2 fold changes,
and in this case is labeled using normal scale fold change
values. The threshold is defined with \code{fold_cutoff} using
normal space values. The log2 fold change values which
have greater magnitude than \code{fold_cutoff} are marked
"changing".

Significance usually represents adjusted P-value, or raw
P-value if necessary. The threshold is defined with \code{sig_cutoff}
using a P-value below which entries are marked "significant".

Finally, since some statistical criteria also include a minimum
level of signal, a threshold \code{expr_cutoff} requires an entry to
have signal at or above this value to be considered "detected".

The default behavior of \code{volcano_plot()} is to render a
smooth scatter plot. A smooth scatter plot is much more
effective at representing the true point density along
the figure, which is one of the primary reasons to produce
the plot.
\subsection{Highlighting points}{

The argument \code{hi_points} can be used to highlight a specific
subset of points on the figure, even when \code{smooth=TRUE}.

Alternatively, \code{hi_hits=TRUE} will render all statistical
hits as points, which will appear on top of the smooth
scatter plot when \code{smooth=TRUE}.
}
}
\examples{
n <- 15000;
set.seed(12);
x_lfc <- (rnorm(n) * 1);
x_lfc <- x_lfc^2 * sign(x_lfc);
x_lfc <- x_lfc[order(-abs(x_lfc) + rnorm(n) / 2)];
x_pv <- sort(10^-(rnorm(n)*1.5)^2);
x <- data.frame(
   Gene=paste("gene", seq_len(n)),
   `log2fold Group-Control`=x_lfc,
   `P.Value Group-Control`=x_pv[order(-abs(x_lfc))],
   `mgm Group-Contol`=((rnorm(1500)+5)^2)/5,
   check.names=FALSE);

vdf <- volcano_plot(x);
volcano_plot(x, expr_cutoff=3);
# volcano_plot(x, mar_min=c(7, 6, 6, 5), blockarrow_font=2, blockarrow_cex=1.5);

# par("mfrow"=c(2, 1));
# volcano_plot(x);
# volcano_plot(x);
# par("mfrow"=c(1, 1));

x[["fold Group-Control"]] <- log2fold_to_fold(x[["log2fold Group-Control"]]);
x[["adj.P.Val Group-Control"]] <- x[["P.Value Group-Control"]];

volcano_plot(x, hi_hits=TRUE);

volcano_plot(x, hi_hits=TRUE, label_hits=TRUE)

}
\seealso{
Other jam plot functions:
\code{\link[=ggjammaplot]{ggjammaplot()}},
\code{\link[=ggvolcano_plot]{ggvolcano_plot()}}
}
\concept{jam plot functions}
