% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamma-volcano.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{Create Volcano plot of a statistical comparison}
\usage{
volcanoPlot(x, n = NULL, coef = NULL,
  geneColumn = head(provigrep(geneColGrep, colnames(x)), 1),
  lfcCol = head(provigrep(lfcColGrep, colnames(x)), 1),
  pvCol = head(provigrep(pvColGrep, colnames(x)), 1),
  intensityCols = provigrep(intensityColGrep, colnames(x)),
  geneColGrep = c("^gene", "symbol"), lfcColGrep = c("log.{0,2}fc", "lfc",
  "log.{0,4}fold", "fold.*ch", "fc", "fold", "log.*ratio", "ratio"),
  pvColGrep = c("^padj|adj.*p", "fdr", "q.*val", "p.*val"),
  intensityColGrep = c("maxgroupmean|maxmean", "groupMean", "aveexpr",
  "^fpkm"), pvalueCutoff = 0.01, lfcCutoff = 0, intensityCutoff = NULL,
  minPvalue = 1e-20, pvalueFloor = minPvalue, maxLog2FC = 10,
  xRange = NULL, yRange = NULL, lfcCeiling = NULL, margins = c(5, 6, 4,
  2), main = "Volcano Plot", lineMain = 3, subMain = NULL,
  font.main = 1, symmetricAxes = TRUE, xlab = paste("Fold change (",
  lfcCol, ")", sep = ""), ylab = paste("Significance (", pvCol, ")", sep =
  ""), nXlabels = 13, nYlabels = 7, ablineColor = "#000000AA",
  orientation = "up", addPlot = FALSE, doStatsSubtitles = TRUE,
  statsLine = 4, statsCex = 0.8, ptCex = 0.9, ptPch = 21,
  hitType = "hits", hitColorSet = c("#77777777", "#99000055"),
  pointColors = NULL, pointBgColors = NULL, pointColorSet = c(base =
  "#77777733", up = "#99000055", down = "#00009955", base.highlight =
  "#FFDD55DD", up.highlight = "#FFDD55DD", down.highlight = "#FFDD55DD"),
  pointBorderSet = c(base = "#33333333", up = "#77000088", down = "#00007788",
  base.highlight = "#773322DD", up.highlight = "#773322DD", down.highlight =
  "#773322DD"), contrastFactor = 1.5, highlightPoints = NULL,
  highlightHits = FALSE, highlightItemType = paste("highlighted", hitType,
  sep = " "), highlightCex = 1, pointsToLabel = NULL,
  noObscureMode = TRUE, initialAngle = 0, initialRadius = 0,
  labelCex = 1, labelMar = labelCex * 1.1, boxColor = "#EEBB7788",
  boxBorderColor = "#000000AA", labelFixedCoords = NULL,
  doHighlightLabels = TRUE, sectionLabelSpacing = 12, labelN = 30,
  doJitter = 10, jitterAmount = NULL, hitsByGene = FALSE,
  printMultiGeneHits = FALSE, doSmoothScatter = TRUE,
  smoothScatterFunc = plotSmoothScatter, colramp = c("white", "lightblue",
  "lightskyblue3", "royalblue", "darkblue", "orange", "darkorange1",
  "orangered2"), useRaster = TRUE, transFactor = 0.24,
  transformation = function(x) {     x^transFactor }, doBoth = FALSE,
  bothColorOnlyHits = TRUE, pointSubset = NULL, plotOnlySubset = FALSE,
  nbin = 256, nrpoints = 0, doBlockArrows = TRUE, blockArrowFont = 1,
  blockArrowHitColor = "#E67739FF", blockArrowUpColor = "#990000FF",
  blockArrowDownColor = "#000099FF", blockArrowLabelHitColor = "#FFFFAA",
  blockArrowLabelUpColor = "#FFFFAA", blockArrowLabelDownColor = "#FFFFAA",
  blockArrowCex = c(1, 1), blockArrowLabelCex = 1,
  doShadowTextArrows = TRUE, doTopHist = FALSE,
  doTopHistCutoffs = c("pvalue", "foldchange"), topHistBreaks = 100,
  topHistColor = "#000099FF", topHistCexSub = 1, topHistSubLine = 0,
  topHistPlotFraction = 1/5, topHistBy = 0.2, cexSub = 0.8,
  cexMain = 1.5, cex.axis = 1.2, lineXlab = 2.5, lineYlab = 4,
  labelHits = !doBlockArrows, verbose = FALSE, ...)
}
\description{
Create Volcano plot of a statistical comparison
}
\details{
This function takes a data.frame of statistical results in the form
of contrasts with fold changes and P-values, and produces a volcano
plot with -log10 P-value on the y-axis, versus log2 fold change on
the x-axis. The primary purpose of the plot is to observe the distribution
of up- and down- changes, the distribution of P-values, relative to
optionally supplied statistical cutoffs used to define statistical hits.

To do:

\itemize{
\item{Adjust y-axis margin so the y-axis label is displayed, since it
indicates on the plot which P-value column is being displayed.}
\item{Aesthetics, Increase default bandwidthN. Increase default
ylab and xlab font sizes. Consider higher contrast color for the
y-axis statistical hits block arrow.}
\item{Include hits with returned data, it should be easy to retrieve
the exact entries marked as hits on the plot.}
}

Note: This function is not exported, as it is still being ported
to the Jam way of doing things.
}
