% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-volcano-plot.R
\name{blockArrowMargin}
\alias{blockArrowMargin}
\title{Draw block arrows in plot margins}
\usage{
blockArrowMargin(
  axisPosition = "rightAxis",
  arrowPosition = "top",
  arrowLabel = "",
  arrowDirection = "updown",
  xleft = NULL,
  xright = NULL,
  ybottom = NULL,
  ytop = NULL,
  col = "#660000FF",
  labelFont = 1,
  labelCex = 1,
  border = "#000000FF",
  xpd = TRUE,
  parUsr = par("usr"),
  arrowWidthPercent = 6,
  blockWidthPercent = 5,
  arrowLengthPercent = blockWidthPercent * 0.6,
  bufferPercent = 0.5,
  blankFirst = FALSE,
  arrowLabelColor = NULL,
  arrowLabelBorder = "#000000FF",
  doExample = FALSE,
  doBlockGradient = TRUE,
  doShadowText = TRUE,
  gradientDarkFactor = 1.5,
  gradientSFactor = 1.5,
  verbose = FALSE,
  ...
)
}
\description{
Draw block arrows in plot margins
}
\details{
This function draws block arrows in plot margins,
intended to be used to describe plot axis ranges
with an optional label. The driving example is to
describe the number of statistical hits shown
on a volcano plot.

Run \code{blockArrowMargin(doExample=TRUE)} to see a visual
example.
}
\examples{
blockArrowMargin(doExample=TRUE)

}
\seealso{
Other jam utility functions:
\code{\link[=find_colname]{find_colname()}},
\code{\link[=fold_to_log2fold]{fold_to_log2fold()}},
\code{\link[=get_se_assaydata]{get_se_assaydata()}},
\code{\link[=get_se_colData]{get_se_colData()}},
\code{\link[=get_se_rowData]{get_se_rowData()}},
\code{\link[=gradient_rect]{gradient_rect()}},
\code{\link[=handle_highlightPoints]{handle_highlightPoints()}},
\code{\link[=log2fold_to_fold]{log2fold_to_fold()}},
\code{\link[=logAxis]{logAxis()}},
\code{\link[=outer_legend]{outer_legend()}},
\code{\link[=points2polygonHull]{points2polygonHull()}},
\code{\link[=update_function_params]{update_function_params()}},
\code{\link[=update_list_elements]{update_list_elements()}}
}
\concept{jam utility functions}
