% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-element_textbox_colorsub.R
\name{element_grob.element_textbox_colorsub}
\alias{element_grob.element_textbox_colorsub}
\title{Custom ggplot2 element textbox highlight grob}
\usage{
element_grob.element_textbox_colorsub(element, label = "", ...)
}
\description{
Custom ggplot2 element textbox highlight grob
}
