% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamma-se-data.R
\name{get_se_assaydata}
\alias{get_se_assaydata}
\title{Get SummarizedExperiment assay matrix data}
\usage{
get_se_assaydata(x, assay_name = NULL, verbose = FALSE, ...)
}
\arguments{
\item{x}{\code{SummarizedExperiment} object}

\item{assay_name}{\code{character} string that should match one entry
in \code{names(assays(x))}.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\description{
Get SummarizedExperiment assay matrix data
}
\seealso{
Other jam utility functions: 
\code{\link{blockArrowMargin}()},
\code{\link{find_colname}()},
\code{\link{fold_to_log2fold}()},
\code{\link{get_se_colData}()},
\code{\link{gradient_rect}()},
\code{\link{handle_highlightPoints}()},
\code{\link{log2fold_to_fold}()},
\code{\link{logAxis}()},
\code{\link{outer_legend}()},
\code{\link{points2polygonHull}()},
\code{\link{update_function_params}()},
\code{\link{update_list_elements}()}
}
\concept{jam utility functions}
