% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamma-ggjammaplot.R
\name{get_se_assaydata}
\alias{get_se_assaydata}
\title{Get SummarizedExperiment assay matrix data}
\usage{
get_se_assaydata(x, assay_name = NULL, verbose = FALSE, ...)
}
\arguments{
\item{x}{\code{SummarizedExperiment} object}

\item{assay_name}{\code{character} string that should match one entry
in \code{names(assays(x))}.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\description{
Get SummarizedExperiment assay matrix data
}
\seealso{
Other jam utility functions: 
\code{\link{handle_highlightPoints}()},
\code{\link{outer_legend}()},
\code{\link{points2polygonHull}()}
}
\concept{jam utility functions}
