% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamma-ggjammaplot.R
\name{get_se_assaydata}
\alias{get_se_assaydata}
\title{Get SummarizedExperiment assay matrix data}
\usage{
get_se_assaydata(x, assay_name = NULL, verbose = FALSE, ...)
}
\description{
Get SummarizedExperiment assay matrix data
}
\seealso{
Other jam utility functions: 
\code{\link{element_grob.element_textbox_colorsub}()},
\code{\link{element_textbox_colorsub}()},
\code{\link{handle_highlightPoints}()},
\code{\link{outer_legend}()},
\code{\link{points2polygonHull}()}
}
\concept{jam utility functions}
