% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-gg-volcano-plot.R
\name{ggvolcano_plot}
\alias{ggvolcano_plot}
\title{Draw a volcano plot using ggplot2 with reasonable default arguments.}
\usage{
ggvolcano_plot(
  x,
  base_size = 20,
  fold_cutoff = 1.5,
  sig_cutoff = 0.05,
  cutoff_color = "#00000077",
  cutoff_linetype = "dashed",
  cutoff_linewidth = 1,
  fold_style = c("fold", "log2fold", "log2", "lfc"),
  baseColor = "white",
  panel.grid.major.colour = "grey90",
  panel.grid.minor.colour = "transparent",
  axis.text.x.angle = 60,
  titleCex = 1,
  detail_factor = 1,
  nbin = 300,
  bw_factor = 1,
  transFactor = 0.24,
  smooth_colors = c("transparent", "lightblue", "lightskyblue3", "royalblue", "darkblue",
    "orange", "darkorange1", "orangered2"),
  aspect_adjust = 1,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{\code{data.frame} with data suitable for volcano plot.}

\item{base_size}{\code{numeric} base font point size used for the
ggplot2 theme \code{colorjam::theme_jam()}, default 20 intended
for presentation or publication figures.}

\item{panel.grid.major.colour, panel.grid.major.colour}{\code{character}
colors used for the major and minor grid lines.}

\item{axis.text.x.angle}{\code{numeric} angle used for the x-axis
label text, default 0.}

\item{titleCex}{\code{numeric} title font adjustment}

\item{detail_factor}{\code{numeric} adjustment to increase overall
detail in the density plot.}

\item{nbin}{\code{integer} number of bins to use for density plot.}

\item{bw_factor}{\code{numeric} adjustment to increase detail used
in calculating the detail in the point density.}

\item{transFactor}{\code{numeric} exponent value used to adjust the
point density, similar to that used by \code{smoothScatter()}
in the form \code{function(x)x^transFactor}.}

\item{smooth_colors}{\code{character} vector of colors, or single color
or single color ramp name recognized by \code{jamba::getColorRamp()}.}

\item{...}{additional arguments are passed to \code{volcano_plot()}.}
}
\value{
\code{ggplot2} graphics object, where the \code{data.frame}
is obtained with \code{gg@data}.
}
\description{
Draw a volcano plot using ggplot2 with reasonable default
arguments, and with a large number of customization options.
The default plot uses smooth scatter plot for much improved
display of point density.
}
\details{
See \code{volcano_plot()} for the detailed customization options.

Note that ggplot2 does not currently display block arrows with
the number of hits that met statistical thresholds.
\subsection{To label genes}{
\itemize{
\item Define \code{hi_points} for the points to be highlighted.
\itemize{
\item It can be a \code{list} of \code{character} vectors, which are
matched with gene values in \code{gene_colname}.
The points are colored by \code{hi_colors} which should be
a \code{character} vector of colors the same length as
\code{hi_points} list length, optionally named using
\code{names(hi_points)},
\item It can be a \code{character} vector of gene values.
This input is treated as a \code{list} containing one
\code{character} vector.
\item \strong{The color legend is not yet implemented}
for \code{ggvolcano_plot()}.
}
\item Define \code{hi_do_label} with \code{logical} vector, \code{character}
vector, or \code{integer} index.
\itemize{
\item Labels use the column matched by \code{label_colname}, which
by default is the same column as \code{gene_colname}.
\item The label color is inherited from \code{hi_colors} except it
it made much lighter, and desaturated, using
\code{jamba::makeColorDarker(..., darkFactor=-1.6, sFactor=-1.5)}
}
}
}
}
\examples{
n <- 15000;
set.seed(12);
x_lfc <- (rnorm(n) * 1);
x_lfc <- x_lfc^2 * sign(x_lfc);
x_lfc <- x_lfc[order(-abs(x_lfc) + rnorm(n) / 2)];
x_pv <- sort(10^-(rnorm(n)*1.5)^2);
x <- data.frame(
   Gene=paste("gene", seq_len(n)),
   `log2fold Group-Control`=x_lfc,
   `P.Value Group-Control`=x_pv[order(-abs(x_lfc))],
   `mgm Group-Contol`=((rnorm(1500)+5)^2)/5,
   check.names=FALSE);

ggvdf <- ggvolcano_plot(x);
plot(ggvdf)

}
\seealso{
Other jam plot functions: 
\code{\link{ggjammaplot}()},
\code{\link{volcano_plot}()}
}
\concept{jam plot functions}
