% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-volcano-plot.R
\name{gradient_rect}
\alias{gradient_rect}
\title{Rectangle with color gradient fill}
\usage{
gradient_rect(
  xleft,
  ybottom,
  xright,
  ytop,
  col,
  gradient = "x",
  border = par("fg"),
  ...
)
}
\arguments{
\item{xleft, ybottom, xright, ytop}{\code{numeric} vectors indicating
the position of sides of a rectangle, passed to
\code{graphics::rect()}. Multiple rectangles may be defined.}

\item{col}{\code{character} vector of colors used to fill the rectangles.}

\item{gradient}{\code{character} string indicating the direction of
color gradient, with two allowed values: \code{"x"} and \code{"y"}.}

\item{border}{\code{character} value indicating the color of border
around the rectangle.}

\item{...}{additional arguments are ignored.}
}
\description{
Rectangle with color gradient fill
}
\details{
This function was inspired by the \code{plotrix::gradient.rect()}
function in the plotrix R package. The function is
simplified here, and requires a vector of colors in \code{col}.
}
\examples{
jamba::nullPlot(xlim=c(0,5), ylim=c(0,5), xaxt="s", yaxt="s");
gradient_rect(xleft=1,
   ybottom=1,
   xright=2.5,
   ytop=2.5,
   col=jamba::getColorRamp("Reds", n=15))
gradient_rect(xleft=2.5,
   ybottom=2.5,
   xright=4,
   ytop=4,
   gradient="y",
   col=jamba::getColorRamp("Reds", n=15))

}
\seealso{
Other jam utility functions: 
\code{\link{blockArrowMargin}()},
\code{\link{find_colname}()},
\code{\link{fold_to_log2fold}()},
\code{\link{get_se_assaydata}()},
\code{\link{handle_highlightPoints}()},
\code{\link{log2fold_to_fold}()},
\code{\link{logAxis}()},
\code{\link{outer_legend}()},
\code{\link{points2polygonHull}()},
\code{\link{update_function_params}()},
\code{\link{update_list_elements}()}
}
\concept{jam utility functions}
