% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamma-ggjammaplot.R
\name{handle_highlightPoints}
\alias{handle_highlightPoints}
\title{Handle highlightPoints argument to jammaplot()}
\usage{
handle_highlightPoints(
  highlightPoints = NULL,
  highlightColor = NULL,
  highlightPch = 21,
  highlightCex = 1.5,
  ...
)
}
\description{
Handle highlightPoints argument to jammaplot()
}
\seealso{
Other jam utility functions: 
\code{\link{fold_to_log2fold}()},
\code{\link{get_se_assaydata}()},
\code{\link{log2fold_to_fold}()},
\code{\link{outer_legend}()},
\code{\link{points2polygonHull}()}
}
\concept{jam utility functions}
