% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamma-ggjammaplot.R
\name{handle_highlightPoints}
\alias{handle_highlightPoints}
\title{Handle highlightPoints argument to jammaplot()}
\usage{
handle_highlightPoints(
  highlightPoints = NULL,
  highlightColor = NULL,
  highlightPch = 21,
  highlightCex = 1.5,
  ...
)
}
\description{
Handle highlightPoints argument to jammaplot()
}
\seealso{
Other jam utility functions: 
\code{\link{blockArrowMargin}()},
\code{\link{find_colname}()},
\code{\link{fold_to_log2fold}()},
\code{\link{get_se_assaydata}()},
\code{\link{get_se_colData}()},
\code{\link{gradient_rect}()},
\code{\link{log2fold_to_fold}()},
\code{\link{logAxis}()},
\code{\link{outer_legend}()},
\code{\link{points2polygonHull}()},
\code{\link{update_function_params}()},
\code{\link{update_list_elements}()}
}
\concept{jam utility functions}
