% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamma.r
\docType{package}
\name{jamma}
\alias{jamma}
\alias{jamma-package}
\title{jamma: MA-plots for omics data}
\description{
The jamma package creates MA-plots for omics data, with
important customizations to handle specific controls, and
experimental subsets.
}
\section{MA-plots functions}{

jammaplot,
centerGeneData
}

\section{Additional plot functions}{

points2polygonHull
}

