% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamma.r
\name{jammaplot}
\alias{jammaplot}
\alias{MAplot,}
\alias{MVAplotMED,}
\alias{MVAplotMEDsmooth,}
\alias{MVAplot,}
\alias{plotMA,}
\alias{plotMVA,}
\alias{ma.plot,}
\alias{plot.ma,}
\alias{mva.plot}
\title{Produce MA-plot of omics data.}
\usage{
jammaplot(
  x,
  maintitle = NULL,
  titleBoxColor = "#DDBB9977",
  subtitleBoxColor = titleBoxColor,
  centerGroups = NULL,
  controlSamples = colnames(x),
  useMean = TRUE,
  ylim = c(-4, 4),
  xlim = NULL,
  highlightPoints = NULL,
  outlierMAD = 5,
  outlierRowMin = 5,
  displayMAD = FALSE,
  groupedMAD = TRUE,
  colramp = c("white", "lightblue", "blue", "navy", "orange", "orangered2"),
  colrampOutlier = NULL,
  outlierColor = "palegoldenrod",
  whichSamples = NULL,
  maintitleCex = 1.8,
  subtitle = NULL,
  subtitlePreset = "bottomleft",
  titleCexFactor = 1,
  titleCex = NULL,
  doTitleBox = TRUE,
  titleColor = "black",
  titleFont = 2,
  titlePreset = "top",
  xlab = "",
  xlabline = 2,
  ylab = "",
  ylabline = 1.5,
  groupSuffix = NULL,
  highlightPch = 21,
  highlightCex = 1.5,
  highlightColor = "#00AAAA66",
  doHighlightPolygon = FALSE,
  highlightPolygonAlpha = 0.3,
  doHighlightLegend = TRUE,
  smoothPtCol = "#00000055",
  margins = c(3.5, 2, 0.3, 0.2),
  useRaster = TRUE,
  ncol = NULL,
  nrow = NULL,
  doPar = TRUE,
  las = 2,
  groupedX = TRUE,
  customFunc = NULL,
  filterNA = TRUE,
  filterNAreplacement = NA,
  filterNeg = TRUE,
  filterFloor = 0,
  filterFloorReplacement = filterFloor,
  transFactor = 0.18,
  nrpoints = 0,
  smoothScatterFunc = jamba::plotSmoothScatter,
  applyRangeCeiling = TRUE,
  doTxtplot = FALSE,
  ablineV = 0,
  ablineH = c(-2, 0, 2),
  blankPlotPos = NULL,
  fillBackground = TRUE,
  useRank = FALSE,
  ma_method = c("jammacalc", "old"),
  doPlot = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{\code{numeric} \code{matrix} typically containing log-normal
values, with measurement rows, and sample columns. For example,
with gene or protein expression data, the genes or proteins are
represented in rows, and biological samples are represented
in columns.}

\item{maintitle}{\code{character} string with the main title for
the overall plot, printed at the top of each page in the
top outer margin.}

\item{titleBoxColor, subtitleBoxColor}{\code{character} vector of
R colors applied to title text, or subtitle text, respectively.
When \code{doTitleBox=TRUE}
one or no value is supplied, it defines colors using
\code{jamba::setTextContrastColor()} to use a contrasting color.}

\item{centerGroups}{vector of groups, with length \code{ncol(x)}
indicating optional groupings to use during data centering.
When \code{centerGroups} is defined, each subset of data is centered
independently. It is useful to center within batches, or within
subsets of samples that are not intended to be compared across
one another. It is also informative to center by each sample
group in order to view the variability among sample group
replicates, which should be much lower than variability across
sample groups. See \code{centerGeneData()} for more specific examples.}

\item{controlSamples}{\code{character} vector of \code{colnames(x)} to define
control samples during the data centering step. Default MA-plots are
calculated relative to the overall mean, which uses all \code{colnames(x)}
as \code{controlSamples}. However it can be helpful and informative
to view MA-plots relative to a set of known control samples.}

\item{useMean}{\code{logical} indicating whether to center data using
\code{mean} or \code{median} values. The default \code{useMean=TRUE} is chosen
because it visually represents data as seen by typical parametric
analysis methods downstream. When outlier sample are observed,
it may be more useful to apply \code{useMean=FALSE} to use the \code{median}
which is less prone to outlier effects. That said, if a particular
sample is an outlier, another alternative is to define
\code{controlSamples} which excludes the outlier sample(s), and
therefore data centering is applied using only the non-outlier
samples as the reference.}

\item{ylim, xlim}{\code{NULL} or \code{numeric} vector \code{length=2} indicating
the y-axis and x-axis ranges, respectively. The values are useful
to define consistent dimensions across all panels. The
default \code{ylim=c(-4,4)} represents 16-fold up and down range in
normal space, and is typically a reasonable starting point for
most purposes. Even if numeric values are all between
\code{-1.5} and \code{1.5}, it is still recommended to keep a range in
context of \code{c(-4, 4)}, to indicate that the observed values
are lower than typically observed. The \code{c(-4, 4)} may be adjusted
relative to the typical ranges expected for the data.
It is sometimes helpful to define \code{xlim} slightly above zero for
datasets that have an extremely large proportion of zeros, in order
to reduce the visual effect of having that much point density at
zero, for example with \code{xlim=c(0.001, 20)} and
\code{applyRangeCeiling=FALSE}.}

\item{highlightPoints}{\code{NULL}, or \code{character} vector, or a
\code{list} of \code{character} vectors indicating \code{rownames(x)} to
highlight in each MA-plot panel. When \code{NULL}, no points are
highlighted; when \code{character} vector, points are highlighted in
all MA-plot panels; when \code{list} of \code{character} vectors, each
\code{character} vector in the list is highlighted using a unique
color in \code{highlightColor}. Points are drawn using
\code{graphics::points()} and colored using \code{highlightColor},
which can be time-consuming for a large number of highlight
points.}

\item{outlierMAD}{\code{numeric} threshold above which a MA-plot panel
MAD factor is considered an outlier. When a MA-plot panel is
considered an outlier, the \code{outlierColramp} or \code{outlierColor}
is applied to the panel color ramp to display a visual
indication.}

\item{outlierRowMin}{\code{numeric} value indicating the minimum mean
value as displayed on the MA-plot panel x-axis, in order for
the row to be included in MAD calculations. This argument is
intended to prevent measurements whose mean value is below
a noise threshold from being included, therefore only including
points whose mean measurement is above noise and represents
"typical" variability.}

\item{displayMAD}{\code{logical} indicating whether to display each
MA-plot panel MAD factor (median absolute deviation). A MAD
value for each panel is calculated by taking the median absolute
deviation from zero across all points, using points whose mean
value is equal or greater than \code{outlierRowMin}. The overall MAD
is defined by the median MAD from the MA-plot panels. The MAD
factor is defined as the ratio of each MA-plot panel MAD value
to the overall MAD value, and therefore most MAD factor values
should be roughly \code{1}. The overall MAD value is defined by the
median across all samples when \code{groupedMAD=FALSE}, or defined
within each \code{centerGroup} when \code{groupedMAD=TRUE}. A value
with MAD factor 2 is interpreted as a sample whose median
deviation from zero is twice as high as the typical sample,
which is a reasonably indication that this sample has twice
the inherent level of noise compared to other samples. Note
that MAD values should be interpreted within sample processing
batches if relevant, or within logical experimental units --
roughly interpreted to mean sets of samples within which direct
statistical comparisons are intended to be applied. For example,
gene expression data that include brain and liver samples would
probably use \code{centerGroups} for brain and liver to be centered
separately, therefore the MAD factors should be separately
calculated for brain and for liver.}

\item{groupedMAD}{logical indicating how the MAD calculation
should be performed: \code{groupedMAD=TRUE} will calculate the
median MAD and corresponsing MAD factor within each
\code{centerGroups} grouping; \code{groupedMAD=FALSE} will calculate
one overall median MAD, and corresponding MAD factor values
will be performed across all samples.}

\item{colramp}{one of several inputs recognized by
\code{jamba::getColorRamp()}. It typically recognizes either the name of
a color ramp from RColorBrewer, the name of functions from the
\code{viridis} package such as \code{viridis::viridis()}, or single R colors, or
a vector of R colors. When a single color is supplied, a gradient
is created from white to that color, where the default base color
can be customized with \code{defaultBaseColor="black"} for example.}

\item{colrampOutlier}{one of several inputs recognized by
\code{jamba::getColorRamp()} to define a specific color ramp for
MA-plot outlier panels, used when \code{outlierMAD} is defined.
When \code{colrampOutlier} is \code{NULL} the \code{outlierColor} is used.}

\item{outlierColor}{\code{character} string representing one R color,
used when \code{colrampOutlier} is \code{NULL} and when \code{outlierMAD} is
defined. This color is used for MA-plot outlier panels by
substituting the first color from the \code{colramp} color ramp,
to act as a visual cue that the panel represents an outlier.}

\item{whichSamples}{\code{NULL} or \code{integer} vector, representing
an index subset of samples to include in the MA-plots. When
\code{whichSamples} represents a subset of samples in \code{x}, the
MA-plot calculations are performed on all samples, then only
samples in \code{whichSamples} are displayed. This argument keeps
the MA-plot calculations consistent even when viewing only
one or a subset of samples in more detail.}

\item{maintitleCex}{\code{numeric} cex character expansion used to
resize the \code{maintitle}.}

\item{subtitle}{\code{NULL} or \code{character} vector to be drawn at
the bottom left corner of each plot panel, the location
is defined by \code{subtitlePreset}.}

\item{subtitlePreset}{character value describing where to position the
subtitle, using terms valid in \code{jamba::coordPresets()}. The default
\code{subtitlePreset="bottomleft"} places the subtitle at the
bottom left corner of each plot panel.}

\item{doTitleBox}{\code{logical} indicating whether to draw plot
titles using a colored box. When \code{doTitleBox=TRUE} the
\code{jamba::drawLabels()} is called to display a label box at
the top of each plot panel, with \code{drawBox=TRUE}. When
\code{doTitleBox=FALSE}, \code{jamba::drawLabels()} is called with
\code{drawBox=FALSE}.}

\item{titleColor}{\code{character} vector of colors applied to title text
in each MA-plot panel. When \code{doTitleBox=TRUE} and \code{titleColor}
contains only one or no value, the title color is defined by
\code{jamba::setTextContrastColor()} along with \code{titleBoxColor}.}

\item{titleFont}{\code{integer} font compatible with \code{par("font")}.
Values are recycled across panels, so each panel can use a custom
value if needed.}

\item{titlePreset}{\code{character} value describing where to position the
subtitle, using terms valid in \code{jamba::coordPresets()}.
The default \code{titlePreset="top"} centers the label at the
top of each panel.}

\item{xlab, ylab}{\code{character} x- and y-axis labels, respectively.
The default values are blank \code{""} because there are a wide variety
of possible labels, and the labels take up more space
than is often useful for most MA-plots.}

\item{xlabline, ylabline}{\code{numeric} number indicating the text line
distance from the edge of plot border to place \code{xlab} and \code{ylab}
text, as used by \code{graphics::title()}.}

\item{groupSuffix}{\code{character} text appended to each MA-plot
panel title. This argument is deprecated in favor of
using \code{subtitle} to place additional text at the bottom left
corner of each MA-plot panel.}

\item{highlightCex}{\code{numeric} value recycled to \code{length(highlightPoints)}
indicating the highlight point size.}

\item{highlightColor}{\code{character} vector used when \code{highlightPoints}
is defined. It is recycled to \code{length(highlightPoints)} and
is applied either to}

\item{doHighlightPolygon}{\code{logical} indicating whether to draw
a shaded polygon encompassing \code{highlightPoints}, using
\code{highlightColor}.The polygon is defined by \code{grDevices::chull()}
via the function \code{points2polygonHull()}.}

\item{highlightPolygonAlpha}{\code{numeric} value indicating alpha
transparency, where \code{0} is fully transparent, and \code{1} is completely
not transparent.}

\item{doHighlightLegend}{\code{logical} indicating whether to print a
color legend when \code{highlightPoints} is defined. The legend is
displayed in the bottom outer margin of the page using
\code{outer_legend()}, and the page is adjusted to add bottom
outer margin.}

\item{smoothPtCol}{\code{color} used to draw points when \code{nrpoints} is
non-zero, which draws points in the extremities of the
smooth scatter plot. See \code{jamba::plotSmoothScatter()}.
The effect can also be achieved by adjusting \code{transFactor} to
a lower value, which increases the visual contrast of individual
points in the point density.}

\item{margins}{\code{numeric} vector of margins compatible with
\code{graphics::par("mar")}. Default values
are applied, but provided here for convenience.}

\item{useRaster}{\code{logical} indicating whether to draw the
smooth scatter plot using raster logic, \code{useRaster=TRUE} is
passed to \code{jamba::plotSmoothScatter()}. The default \code{TRUE}
creates a much smaller plot object by rendering each plot
panel as a single raster image instead of rendering individual
colored rectangles.}

\item{ncol, nrow}{\code{integer} number of MA-plot panel columns and rows
passed to \code{graphics::par("mfrow")} when \code{doPar=TRUE}. When only one
value is supplied, \code{nrow} or \code{ncol}, the other value is defined
by \code{ncol(x)} and \code{blankPlotPos} so all panels can be contained on
one page. When \code{nrow} and \code{ncol} are defined such that multiple
pages are produced, each page will be annotated with \code{maintitle}
and \code{doHighlightLegend} as relevant.}

\item{doPar}{\code{logical} indicating whether to apply
\code{graphics::par("mfrow")} to define MA-plot panel rows and columns.
When \code{doPar=FALSE} each plot panel is
rendered without adjusting the \code{graphics::par("mfrow")} setting.}

\item{las}{\code{integer} value \code{1} or \code{2} indicating whether axis labels
should be parallel or perpendicular to the axes, respectively.}

\item{customFunc}{\code{NULL} or \code{function} used instead of \code{mean} or
\code{median} during the data centering step to generate a row summary
statistic. It should take \code{matrix} input, and return a \code{numeric} vector
output summarizing each row in \code{x}, to be subtracted from each
\code{numeric} value by row in \code{x}. It is intended to
provide custom row statistics, for example geometric mean, or other
row summary function.}

\item{filterNA, filterNAreplacement}{\code{logical} and \code{vector} respectively.
When \code{filterNA=TRUE}, all \code{NA} values are replaced with
\code{filterNAreplacement}, which can be helpful to handle \code{NA} values
as zero \code{0} for example. In reality, \code{NA} values should probably
be left as-is, so subsequent data centering does not use these values,
and so the MA-plot panel does not draw a point when no measurement
exists.}

\item{filterNeg}{\code{logical} deprecated argument, use \code{filterFloor}
instead. The \code{filterNeg} indicates whether to change all negative
values to zero \code{0} before proceeding with data centering.
Negative values are often the result of measurements
being below a noise threshold in upstream data processing,
and therefore the magnitude of negative value is usually
either not informative, or not on similar scale as positive
values.
When \code{filterNeg=TRUE}, negative values are set to zero, and
can result in a characteristic 45 degree angle line originating at
\code{x=0} extending to the right.}

\item{filterFloor, filterFloorReplacement}{\code{numeric} or \code{NULL} indicating
a numeric floor, where values in \code{x} \strong{at or below} \code{filterFloor} are
replaced with \code{filterFloorReplacement}. Note that this argument
can be used to replace zero \code{0} with \code{NA} in the event that
zeros do not represent measurements. One can typically tell whether
input data includes zero \code{0} values by the presence of characteristic
45-degree angle lines originating from \code{x=0} angled to the right.
The default values replace any values \strong{at or below zero} with zero,
which also applies a numeric floor to negative values.
For some platform data technologies, it can be useful to define a
\code{filterFloor} roughly equivalent to its noise threshold. For example
quantitative PCR sometimes uses \code{log_expression = (40 - Ct)}, where
\code{Ct} values above \code{35} are considered to be noise. That noise threshold
implies that any \code{expression} values \code{5} or lower are roughly
equivalent noise, so applying \code{filterFloor=5} is appropriate.}

\item{transFactor}{\code{numeric} adjustment to the visual density of
points by \code{jamba::plotSmoothScatter()}. The value is based upon
\code{graphics::smoothScatter()} argument \code{transformation} which uses
\code{function(x)x^0.24}. The \code{transFactor} is equivalent to the
exponential in the form: \code{function(x)x^transFactor}. Lower values
increase the visual density more intense, higher values make the
visual density less intense.}

\item{nrpoints}{\code{integer} or \code{NULL} indicating the number of points
to display on the extremity of the smooth scatter density,
passed to \code{jamba::plotSmoothScatter()}.}

\item{smoothScatterFunc}{\code{function} used for the smooth scatter plot,
default \code{jamba::plotSmoothScatter()}. Note that a custom function
may not recognize \code{nrpoints} or \code{transformation}.}

\item{applyRangeCeiling}{\code{logical} passed to
\code{jamba::plotSmoothScatter()} which determines how to handle points
outside the plot x-axis and y-axis range: \code{applyRangeCeiling=TRUE}
will place points at the border of the plot, which is helpful
to indicate that there are more points outside the viewing range;
\code{applyRangeCeiling=FALSE} will crop and remove points outside
the viewing range, which is helpful for example when a large
number of points are at zero and overwhelm the point density.
When there are a large proportion of values at zero, it
can be helpful to apply \code{xlim=c(0.01, 20)} and
\code{applyRangeCeiling=FALSE}.}

\item{doTxtplot}{\code{logical} not yet implemented, indicating whether
to plot results using colored text output.}

\item{ablineH, ablineV}{\code{numeric} vector indicating horizontal and
vertical lines to draw in each MA-plot panel, respectively.
These values are passed to \code{graphics::abline()}.}

\item{blankPlotPos}{\code{NULL} or \code{integer} vector indicating
plot panel positions to be drawn blank, skipped. Blank panel
positions are intended to help customize the visual alignment
of MA-plot panels. The mechanism is similar to \code{ggplot2::facet_wrap()}
except that blank positions can be manually defined by what makes
sense to the experiment design.}

\item{fillBackground}{\code{logical} passed to \code{jamba::plotSmoothScatter()}
indicating whether to fill the plot panel with the background color,
using the first value in the color ramp for each MA-plot panel.}

\item{useRank}{\code{logical} indicating whether to create column-wide
ranks, then create MA-plots using the rank data. When \code{useRank=TRUE}
the y-axis represents the rank difference from mean, and the
x-axis represents the mean rank. Using \code{useRank=TRUE} is a good
method to evaluate whether data can be normalized, or whether
data across samples is inherently noisy.}

\item{ma_method}{character string indicating whether to perform
MA-plot calculations using the old method \code{"old"}; or \code{"jammacalc"}
which uses the function \code{jammacalc()}.}

\item{doPlot}{\code{logical} indicating whether to create plots. When
\code{doPlot=FALSE} only the MA-plot panel data is returned.}

\item{verbose}{logical indicating whether to print verbose output.}

\item{...}{additional parameters sent to downstream functions,
\code{\link{jamba::plotSmoothScatter}}, \code{\link{centerGeneData}}.}
}
\value{
List of data used by jamma() to produce an MA-plot, sufficient
to use as input to produce another MA-plot.
}
\description{
Produce MA-plot of omics data.
}
\details{
\code{jammaplot} takes a numeric matrix, typically of gene expression data,
and produces an MA-plot (Bland-Altman plot), also known as a
median-difference plot. One panel is created for each column of
data. Within each panel, the x-axis represents the mean or median
expression of each row; the y-axis represents the difference from
mean or median for that column.

By default, the plot uses \code{jamba::plotSmoothScatter()}, with optional
highlighted points draw using \code{points()}.

The function will determine an appropriate layout of plot panels,
which can be overridden using \code{ncol} and \code{nrow} to specify the
number of columns and rows of plot panels, respectively. For now,
this function uses base R graphics instead of ggplot2, in order
to accomodate some custom features.

This function uses "useRaster=TRUE" by default, which causes
\code{jamba::plotSmoothScatter()} to render a rasterized image as opposed
to a composite of colored rectangles. This process substantially
reduces the render time in all cases, and reduces the image size
when saving as PDF or SVG.
\subsection{Notable features}{
\subsection{Highlighting points}{

Specific points can be highlighted with argument \code{highlightPoints}
which can be a vector or named list of vectors, containing \code{rownames(x)}.
When using a list, point colors are assigned to each element in the
list in order, using the argument \code{highlightColor}.
}

\subsection{Centering by control samples}{

Typical MA-plots are "global-centered", which calculates the
mean/median across all columns in \code{x}, and this value is subtracted
from each individual value per row.

By specifying \code{controlSamples}
the mean/median is calculated using only the \code{colnames(x)} which match
\code{controlSamples}, thus representing "difference from control."

It may also be useful to center data by known high-quality samples,
so the effect of potential outlier samples is avoided.
}

\subsection{Centering within subgroups}{

By specifying \code{centerGroups} as a vector of group names,
the centering is calculated within each group of \code{colnames(x)}.
In this way, subsets of samples can be treated independently in
the MA-plots. A good example might be producing MA-plots for
"kidney" samples, and "muscle" samples, which may have
fundamentally different signal distributions. A good rule
of thumb is to apply \code{centerGroups} to represent separate
groups of samples where you do not intend to apply direct
statistical comparisons across those samples, without at
least applying a two-way contrast, a fold change of fold
changes.

Another informative technique is to center by sample group,
for example \code{centerGroups=sample_group}.
This technique produces MA-plots that depict the
"difference from group" for each sample replicate of a sample
group, and is very useful for identifying sample replicates
with markedly higher variability to its sample group than
others. In general, the variability within sample group
should be substantially lower than variability across
sample groups. Use \code{displayMAD=TRUE} and \code{outlierMAD=2}
as a recommended starting point for this technique.
}

\subsection{Applying a noise floor}{

The argument \code{filterFloor} provides a numeric lower threshold,
where individual values \strong{at or below} this threshold are
set to a defined value. The default defined value is the
floor itself, which has the effect of removing information from
points that are already below the noise threshold and therefore
are unreliable for this purpose.

Another useful alternative
is to define \code{filterFloor=0} and \code{filterFloorReplacement=NA}
so that values of zero \code{0} are set to \code{NA} and are not included
in the MA-plot calculations, and are not represented as points
in each MA-plot panel. For data with many sparse missing values
represented as zero, these options can be very helpful because
each MA-plot panel will only represent actual measurements,
compared to only those sample which also have actual
measurements for those rows.
}

\subsection{Customizing the panel layout}{

Panels are drawn using the order of \code{colnames(x)} by row,
from left-to-right, then top-to-bottom.
The argument \code{blankPlotPos} is intended to insert an empty panel
at a particular panel position, to help customize the alignment
of sample panels.
This option is typically used with \code{ncol} and \code{nrow} to define
a fixed layout of panel columns and rows. \code{blankPlotPos} refers
to panels numbered as drawn per row of panels,
}

\subsection{Identifying potential sample outliers}{

Use argument \code{displayMAD=TRUE} to display the per-sample MAD factor
relative to its \code{centerGroups} value, if provided. The MAD value
for each MA-plot panel is calculated using rows whose mean
is at or above \code{outlierRowMin}. The median MAD value is calculated
for each \code{centerGroups} grouping when \code{groupedMAD=TRUE}, by default.
Finally, each MA-plot panel MAD factor is the ratio of its MAD value
to the relevant median MAD value. MA-plot panels with MAD factor
above \code{outlierMAD} are considered outliers, and the color ramp
uses \code{outlierColramp} or \code{outlierColor} as a visual cue.

Putative outlier samples should usually not be determined
when:
\itemize{
\item \code{controlSamples} are defined to include only a subset
of sample groups,
\item \code{centerGroups} is not defined, or represents more than one
set of sample groups that are not intended to be statistically
compared directly to one another.
}

Putative outlier samples may be defined when:
\itemize{
\item \code{centerGroups} represents a set of sample groups that are
intended to be involved in direct comparisons
\item \code{centerGroups} represents each sample group
}

Potential sample outliers may be identified by setting a threshold
with \code{outlierMAD}, by default 5xMAD. For a sample to be considered
an outlier, its median difference from mean/median needs to be
five times higher than the median across samples.

We typically recommend an \code{outlierMAD=2} when centering
by sample groups, or when centering within experiment subsets.
For one sample to have 2xMAD factor, its variance needs
to be uniquely twice as high as the majority of other samples, which
is typically symptomatic of possible technical failure.

There are exceptions to this suggested guideline, which includes
scenarios where a batch effect may be involved.

To do:
\itemize{
\item{Accept other object types as input, including Bioconductor
classes: ExpressionSet, SummarizedExperiment, MultiExperimentSet.}
\item{Make it efficient to convey group information, for example
define titleBoxColor with group colors, allow centerByGroup==TRUE
which would re-use known sample group information.'}
\item{Adjust the suffix to indicate when \code{centerGroups} are being
used. For example indicate 'sampleID vs groupA' instead of
'sampleID vs median'.}
}
}

}
}
\examples{
# Note the example data requires the affydata Bioconductor package
if (suppressPackageStartupMessages(require(affydata))) {
   data(Dilution);
   edata <- log2(1+exprs(Dilution));
   jammaplot(edata);
}

}
\seealso{
\code{jamba::plotSmoothScatter()} for enhanced
\code{graphics::smoothScatter()}, \code{centerGeneData()},
\code{jamba::imageDefault()}, \code{jamba::nullPlot()}

Other jam plot functions: 
\code{\link{points2polygonHull}()}
}
\concept{jam plot functions}
