% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamma.r
\name{jammaplot}
\alias{jammaplot}
\alias{MAplot,}
\alias{MVAplotMED,}
\alias{MVAplotMEDsmooth,}
\alias{MVAplot,}
\alias{plotMA,}
\alias{plotMVA,}
\alias{ma.plot,}
\alias{plot.ma,}
\alias{mva.plot}
\title{Produce MA-plot of omics data.}
\usage{
jammaplot(x, colramp = c("white", "lightblue", "blue", "navy", "orange",
  "orangered2"), colrampOutlier = NULL, outlierColor = "palegoldenrod",
  whichSamples = NULL, maintitle = NULL, maintitleCex = 1.8,
  subtitle = NULL, subtitlePreset = "bottomleft", titleCexFactor = 1,
  titleCex = NULL, doTitleBox = TRUE, titleBoxColor = "#DDBB9977",
  subtitleBoxColor = titleBoxColor, titleColor = "black",
  titleFont = 2, titlePreset = "top", xlab = "", xlabline = 2,
  ylab = "", ylabline = 1.5, groupSuffix = NULL,
  highlightPoints = NULL, highlightPch = 21, highlightCex = 1.5,
  highlightColor = "#00AAAA66", doHighlightPolygon = FALSE,
  highlightPolygonAlpha = 0.3, doHighlightLegend = TRUE,
  smoothPtCol = "#00000055", margins = c(3.5, 2, 0.3, 0.2),
  useRaster = TRUE, ncol = NULL, nrow = NULL, doPar = TRUE,
  las = 2, ylim = c(-4, 4), xlim = NULL,
  controlSamples = colnames(x), centerGroups = NULL, groupedX = TRUE,
  useMean = FALSE, customFunc = NULL, filterNA = TRUE,
  filterNAreplacement = NA, filterNeg = TRUE, filterFloor = 0,
  filterFloorReplacement = filterFloor, transFactor = 0.18,
  nrpoints = 50, smoothScatterFunc = jamba::plotSmoothScatter,
  applyRangeCeiling = TRUE, doTxtplot = FALSE, ablineV = 0,
  ablineH = c(-2, 0, 2), blankPlotPos = NULL, outlierMAD = 5,
  outlierRowMin = 5, groupedMAD = TRUE, displayMAD = FALSE,
  fillBackground = TRUE, ma_method = c("old", "jammacalc"),
  doPlot = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{x}{numeric matrix typically containing log-normal measurements,
with measurement rows, and sample columns.}

\item{colramp}{one of several inputs recognized by
\code{\link{getColorRamp}}. It typically recognizes either the name of
a color ramp from RColorBrewer, the name of functions from the
\code{\link[viridis]{viridis}} package, or single R colors, or
a vector of R colors.}

\item{colrampOutlier}{one of several inputs recognized by
\code{\link{getColorRamp}}, to be used only in panels determined to be
outliers based upon MAD outlier logic. Typically the color ramp is
identical to \code{colramp} except the background color indicates
an error, for example yellow background instead of white.}

\item{outlierColor}{character R color, used when \code{colrampOutlier is NULL}
to substitute as the first color from \code{colramp}. This method keeps the
color ramp consistent, but changes the background color to highlight
the plot panel as an outlier.}

\item{whichSamples}{NULL or integer vector, representing an index of samples
to include in the MA-plots. Note that all samples are used to create
the MA-plot data, but only these samples will be plotted, which is useful
when wanting to view only a subset of samples in detail, but where the
MA-plot data is identical to the full dataset.}

\item{maintitle}{the main title for the overall plot, printed at the
top of each page in the top outer margin.}

\item{maintitleCex}{numeric cex character expansion used to resize the
\code{maintitle} when supplied.}

\item{subtitle}{NULL or character text to be drawn at the bottom center
of each plot panel.}

\item{subtitlePreset}{character value describing where to position the
subtitle, using terms valid in \code{jamba::coordPresets()}. The default
\code{subtitlePreset="bottomleft"} places the subtitle at the
bottom left corner of each plot panel.}

\item{doTitleBox}{logical whether to draw plot titles using
a colored box. When \code{doTitleBox=TRUE} the \code{jamba::drawLabels()}
is called to display a label box at the top of each plot panel.}

\item{titleBoxColor, subtitleBoxColor}{vector of colors applied to title
text, or subtitle text, respectively. When \code{doTitleBox=TRUE}
one or no value is supplied, it defines colors using
\code{jamba::setTextContrastColor} to use a contrasting color.}

\item{titleColor}{vector of colors applied to title text. When
\code{doTitleBox=TRUE} one or no value is supplied, it defines
colors using \code{jamba::setTextContrastColor()} to use color
that contrasts with \code{titleBoxColor}.}

\item{titleFont}{integer font compatible with \code{par("font")}.
Values are recycled across panels.}

\item{titlePreset}{character value describing where to position the
subtitle, using terms valid in \code{jamba::coordPresets()}.
The default \code{titlePreset="top"} centers the label at the
top of each panel.}

\item{xlab, ylab}{character x- and y-axis labels.}

\item{xlabline, ylabline}{numeric number of text lines as used by
\code{title} to position labels relative to the plot area.}

\item{groupSuffix}{character text appended to each plot panel title.
This argument is being deprecated in favor of using \code{subtitle}
which places additional text at the bottom corner of each
plot panel.}

\item{highlightPoints}{NULL, character vector, or list of character vectors,
containing points to highlight based upon \code{rownames(object)}. If
a list is supplied, then colors from \code{highlightColor} are recycled
to \code{length(highlightPoints)} so that each list element can receive
its own color. This coloring may be useful for coloring several subsets
of rows, for example housekeeper genes, ribosomal genes, or platform
positive- and negative-control genes.}

\item{highlightColor, highlightCex}{values recycled to the length of
highlightPoints. If highlightPoints is a list, then a list or vector
can be supplied.}

\item{doHighlightPolygon}{logical whether to draw a polygon encompassing
highlighted points. If highlightPoints is a list, a polygon is drawn
around each list element subset of points, using the appropriate
\code{highlightColor} for each list element. The polygon is defined by
\code{grDevices::chull()}.}

\item{highlightPolygonAlpha}{numeric value between 0 and 1 indicating the
alpha transparency to use for the polygon color, based upon
\code{highlightColor}.}

\item{doHighlightLegend}{logical indicating whether to print a
color legend at the bottom of each page, when \code{highlightPoints}
is not \code{NULL}.}

\item{smoothPtCol}{color used when nrpoints>0, and
\code{\link{jamba::plotSmoothScatter}} draws this many points in the extremities.}

\item{margins}{vector of margins compatible with \code{par("mar")}. Defaults
are applied, but provided here for convenient override.}

\item{useRaster}{logical whether to define \code{useRaster=TRUE} in
\code{\link{jamba::plotSmoothScatter}}, which ultimately calls
\code{\link{imageDefault}}. When TRUE, it creates a much smaller
plot object, because it returns a raster image instead of a set of
rectangles when rendering the image.}

\item{ncol, nrow}{integer number of columns and rows used by
\code{par("mfrow")} when \code{doPar=TRUE} is defined. These values
are helpful in defining a fixed layout for multiple MA-plot panels.}

\item{doPar}{logical whether \code{par("mfrow")} should be defined, in
order to fit multiple panels in one plot window. Set to FALSE if the plot
layout is already defined, or if plotting one panel per page, for
example.}

\item{las}{integer value 1 or 2, indicating whether axis labels should be
parallel or perpendicular to the axes, respectively. This default value
is defined and provided here for convenience to override it.}

\item{ylim, xlim}{NULL or vector length 2 indicating the y- and x-axis
ranges, respectively. These values are useful to define upfront when
it is best to focus on fixed, consistent ranges across all samples. The
default \code{ylim=c(-4,4)} represents 16-fold range in normal space, and
is typically a reasonable starting point for most purposes. If values are
all between -1.5 and 1.5, then it is still good to keep that range in
context of -4 and 4, to indicate that the observed noise is lower than
typically observed. It is also helpful to use xlim to trim off zeros
for some data where there might be many undetected entries, combined
with \code{applyRangeCeiling=FALSE} which also removed outlier points
from the plot density calculation. The result can notably boost the
detail displayed in the plot range.}

\item{controlSamples}{vector of colnames(object) which should be used
as control samples in the data centering step. Typical MA-plots are
calculated relative to the overall mean, however it can be insightful
to calculate values relative to a known control group.}

\item{centerGroups}{vector of groups, of \code{length(colnames(object))}
indicating optional subgroups to use when performing data centering.
For example it may be appropriate to center samples of one cell type
only relative to other samples of the same cell type. See
\code{\link{centerGeneData}} for more specifics.}

\item{useMean}{logical whether to perform centering using mean or
median. The default median is intended to represent the data without
being skewed by outliers, however it can be informative to use the mean
values when there are few replicates, or when there are no particular
sample outliers.}

\item{customFunc}{an optional function that is used instead of mean or
median. It should take a matrix input, and return a numeric vector
output summarizing each row in \code{object}. It is intended to
provide custom row statistics, for example geometric mean, or other
row summary function.}

\item{filterNA, filterNeg}{logical indicating whether to remove NA or
negative values before creating the MA-plot. It can be useful to
filter negative values, which are often defined as noise in upstream
normalization steps, so they do not impart visible features into the
MA-plot since they are based upon noise. Negative values are set to
zero when filtered, which gives a characteristic 45 degree angle
symmetry above and below zero.}

\item{filterFloor, filterFloorReplacement}{numeric or NULL, when numeric
values in \code{object} are below \code{filterFloor}, they are replaced
with \code{filterFloorReplacement}. This mechanism is an alternative to
\code{filterNeg}, in that it allows defining a filter above or below
zero. For some platform data , it might be useful to define a
\code{filterFloor} roughly equivalent to its noise threshold.}

\item{transFactor}{numeric adjustment used by
\code{\link{jamba::plotSmoothScatter}}. The default is defined, but available
here as a convenient override. The value is used in a power function in
the form \code{transformation=function(x)x^transFactor}. Lower values
make the scatterplot point density more intense.}

\item{nrpoints}{integer or NULL, the number of points to display on the
extremity, as used by \code{\link{jamba::plotSmoothScatter}}.}

\item{smoothScatterFunc}{function used to plot smooth scatter plot, by
default \code{\link{jamba::plotSmoothScatter}}.}

\item{applyRangeCeiling}{see \code{\link{jamba::plotSmoothScatter}} for details,
logical whether to apply a noise floor and ceiling to display points
outside the plot region at the boundaries of the plot. This parameter
is useful to set FALSE when cropping a plot, in order to remove outlier
points from the density calculation. For example when there is a
very large number of zero values, it can be helpful to define
\code{xlim=c(0.1,20), applyRangeCeiling=FALSE}.}

\item{doTxtplot}{logical whether to plot results in colored text output,
currently under development and is still experimental.}

\item{ablineH, ablineV}{numeric vector indicating where to draw indicator
lines, at these horizontal or vertical positions, respectively.}

\item{blankPlotPos}{NULL or integer vector indicating which plot panels
should be used as blank filler positions. These blank panels can be
used to help organize multiple plot panels to indicate sample groups.
A mechanism similar to ggplot2 facets, except with the ability to
insert known blank positions for organization. The facet_grid mechanic
may be used in future, however is not always ideal depending upon the
experiment design, and so for now a simple manual method is provided.}

\item{outlierMAD}{the MAD factor threshold to define a sample an outlier.}

\item{outlierRowMin}{the minimum value as displayed on the x-axis, for
a row in \code{object} to be used in the MAD outlier calculations. This
threshold is intended to define a noise threshold, so that outlier
samples are determined by using rows above the noise threshold. Thus
a sample outlier would need to have greater than five times median
absolute deviation compared to its center group, based only upon
measurements above noise.}

\item{groupedMAD}{logical indicating whether the MAD calculation should
operate only within \code{centerGroups} (\code{groupedMAD=TRUE}) or whether the
MAD calculation is shared across all samples (\code{groupedMAD=FALSE}).}

\item{displayMAD}{logical whether to display MAD (median absolute deviation)
values for each panel. MAD values are defined for each centerGroup if
supplied, or globally if not. The MAD value is CALCULATED by taking the
median deviation from zero as displayed on the y-axis of each MA-plot
panel, then taking the median of those values. Then a MAD factor is
calculated for each MA-plot panel by taking its median devation divided
by the MAD value for its centerGroup. Most samples should therefore be
close to 1. A value of 2 is interpreted as "two times the MAD factor for
its center group." An outlierMAD of 5x is a reasonable cutoff, which
roughly says that sample has 5 times higher median deviation than other
samples in its center group. If all samples are noisy, it would define
a high MAD value, and therefore all samples would be expected to have
similar MAD factors, near 1.}

\item{fillBackground}{logical sent to \code{\link{jamba::plotSmoothScatter}}
indicating whether to fill the plot panel with the background color,
which is typically the first value in the \code{colramp}. Set to TRUE
when axis ranges may be defined which are outside the range of values
for any particular column in \code{object}.}

\item{ma_method}{character string indicating whether to perform
MA-plot calculations using the old method \code{"old"}; or \code{"jammacalc"}
which uses the newer function \code{jammacalc()}.}

\item{verbose}{logical indicating whether to print verbose output.}

\item{...}{additional parameters sent to downstream functions,
\code{\link{jamba::plotSmoothScatter}}, \code{\link{centerGeneData}}.}
}
\value{
List of data used by jamma() to produce an MA-plot, sufficient
to use as input to produce another MA-plot.
}
\description{
Produce MA-plot of omics data.
}
\details{
\code{jammaplot} takes a numeric matrix, typically of gene expression data,
and produces an MA-plot (Bland-Altman plot), also known as a
median-difference plot. One panel is created for each column of
data. Within each panel, the x-axis represents the mean or median
expression of each row; the y-axis represents the difference from
mean or median for that column.

By default, the plot uses \code{jamba::plotSmoothScatter()}, with optional
highlighted points draw using \code{points()}.

The function will determine an appropriate layout of plot panels,
which can be overridden using \code{ncol} and \code{nrow} to specify the
number of columns and rows of plot panels, respectively. For now,
this function uses base R graphics instead of ggplot2, in order
to accomodate some custom features.

This function uses "useRaster=TRUE" by default, which causes
\code{jamba::plotSmoothScatter()} to render a rasterized image as opposed
to a composite of colored rectangles. This process substantially
reduces the render time in all cases, and reduces the image size
when saving as PDF or SVG.Notable featuresHighlighting points

Specific points can be highlighted with argument \code{highlightPoints}
which can be a vector or named list of vectors, containing \code{rownames(x)}.
When using a list, point colors are assigned to each element in the
list in order, using the argument \code{highlightColor}.Centering by control samples

Typical MA-plots are "global-centered", which calculates the
mean/median across all columns in \code{x}, and this value is subtracted
from each individual value per row.

By specifying \code{controlSamples}
the mean/median is calculated using only the \code{colnames(x)} which match
\code{controlSamples}, thus representing "difference from control."

It may also be useful to center data by known high-quality samples,
so the effect of potential outlier samples is avoided.Centering within subgroups

By specifying \code{centerGroups} as a vector of group names,
the centering is calculated within each group of \code{colnames(x)}.
In this way, subsets of samples can be treated independently in
the MA-plots. A good example might be producing MA-plots for
"kidney" samples, and "muscle" samples, which may have
fundamentally different signal distributions.

Another very useful technique is to center by sample group,
for example \code{centerGroups=sample_group}.
This technique produces MA-plots that depict the
"difference from group", and is very useful for visualizing
the variance of replicates in sample groups. Outlier samples
are sometimes very apparent at this step. See \code{displayMAD} and
\code{outlierMAD} below for detection of potential outlier samples.

Note: filterNeg=TRUE will floor log2 values at 0, which has the
effect of adding the "45 degree lines" originating from c(0,0).Customizing the panel layout

Panels are drawn using the order of \code{colnames(x)} by row,
from left-to-right, then top-to-bottom.
The argument \code{blankPlotPos} is intended to insert an empty panel
at a particular panel position, to help customize the alignment
of sample panels.
This option is typically used with \code{ncol} and \code{nrow} to define
a fixed layout of panel columns and rows. \code{blankPlotPos} refers
to panels numbered as drawn per row of panels,Identifying potential sample outliers

The argument \code{displayMAD} will display the per-sample MAD factor
relative to its \code{centerGroups} value, if provided. The MAD threshold
is calculated using values whose mean/median is at least
\code{outlierRowMin}, based upon the median absolute difference (MAD)
from the mean/median. The median MAD is defined as the reference MAD,
and all other MAD factors are multiples of that reference. For example,
1xMAD is equal to the median MAD value, and 2xMAD is two times higher
than the median MAD value.

Potential outlier samples should not be evaluated when supplying
\code{controlSamples} unless those samples represent more than
experimantal control samples.

Potential sample outliers may be identified by setting a threshold
with \code{outlierMAD}, by default 5xMAD. For a sample to be considered
an outlier, its median difference from mean/median needs to be
five times higher than the median across samples.

We typically recommend an \code{outlierMAD=2} when centering
by sample groups, or when centering within experiment subsets.
For one sample to have 2xMAD factor, its variance needs
to be uniquely twice as high as the majority of other samples, which
is typically symptomatic of possible technical failure.

There are exceptions to this suggested guideline, which includes
scenarios where a batch effect may be involved.

To do:
\itemize{
\item{Accept other object types as input, including Bioconductor
classes: ExpressionSet, SummarizedExperiment, MultiExperimentSet.}
\item{Make it efficient to convey group information, for example
define titleBoxColor with group colors, allow centerByGroup==TRUE
which would re-use known sample group information.'}
\item{Adjust the suffix to indicate when \code{centerGroups} are being
used. For example indicate 'sampleID vs groupA' instead of
'sampleID vs median'.}
}
}
\examples{
# Note the example data requires the affydata Bioconductor package
if (suppressPackageStartupMessages(require(affydata))) {
   data(Dilution);
   edata <- log2(1+exprs(Dilution));
   jammaplot(edata);
}

}
\seealso{
\code{jamba::plotSmoothScatter()} for enhanced
\code{graphics::smoothScatter()}, \code{centerGeneData()},
\code{jamba::imageDefault()}, \code{jamba::nullPlot()}

Other jam plot functions: \code{\link{points2polygonHull}}
}
\concept{jam plot functions}
