% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamma-ggjammaplot.R, R/jamma.r
\name{ggjammaplot}
\alias{ggjammaplot}
\alias{jammaplot}
\alias{MAplot,}
\alias{MVAplot,}
\alias{plotMA,}
\alias{plotMVA,}
\alias{ma.plot,}
\alias{plot.ma}
\title{Produce MA-plot of omics data.}
\usage{
ggjammaplot(
  x,
  detail_factor = 1,
  nbin_factor = 1,
  bw_factor = 1,
  assay_name = 1,
  useMedian = FALSE,
  controlSamples = NULL,
  centerGroups = NULL,
  controlFloor = NA,
  naControlAction = c("row", "floor", "min", "na"),
  naControlFloor = 0,
  colramp = c("transparent", "lightblue", "blue", "navy", "orange", "orangered2"),
  groupedX = TRUE,
  grouped_mad = TRUE,
  outlierMAD = 5,
  mad_row_min = 4,
  displayMAD = FALSE,
  noise_floor = 0,
  noise_floor_value = NA,
  naValue = NA,
  centerFunc = centerGeneData,
  whichSamples = NULL,
  useRank = FALSE,
  apply_transform_limit = 40,
  titleBoxColor = "lightgoldenrod1",
  titleCex = 1,
  subtitleBoxColor = titleBoxColor,
  outlierColor = "lemonchiffon",
  fillBackground = TRUE,
  maintitle = NULL,
  subtitle = NULL,
  summary = "Mean",
  difference = "Difference",
  transFactor = 0.25,
  doPlot = TRUE,
  highlightPoints = NULL,
  highlightPch = 21,
  highlightCex = 1.5,
  highlightColor = NULL,
  doHighlightLegend = TRUE,
  ablineH = c(-2, 0, 2),
  base_size = 12,
  panel.grid.major.colour = "grey90",
  panel.grid.minor.colour = "grey95",
  axis.text.x.angle = 90,
  return_type = c("ggplot", "data"),
  xlim = NULL,
  ylim = c(-6, 6),
  ncol = NULL,
  nrow = NULL,
  blankPlotPos = NULL,
  verbose = FALSE,
  ...
)

jammaplot(
  x,
  assay_name = NULL,
  maintitle = NULL,
  titleBoxColor = "#DDBB9977",
  subtitleBoxColor = titleBoxColor,
  centerGroups = NULL,
  controlSamples = colnames(x),
  controlFloor = NA,
  naControlAction = c("row", "floor", "min", "na"),
  naControlFloor = 0,
  controlIndicator = c("labelstar", "titlestar", "none"),
  sample_labels = NULL,
  useMedian = FALSE,
  useMean = NULL,
  ylim = c(-4, 4),
  xlim = NULL,
  highlightPoints = NULL,
  outlierMAD = 5,
  outlierRowMin = 5,
  displayMAD = FALSE,
  groupedMAD = TRUE,
  colramp = c("white", "lightblue", "blue", "navy", "orange", "orangered2"),
  colrampOutlier = NULL,
  outlierColor = "lemonchiffon",
  whichSamples = NULL,
  maintitleCex = 1.8,
  subtitle = NULL,
  subtitlePreset = "bottomleft",
  subtitleAdjPreset = "topright",
  titleCexFactor = 1,
  titleCex = NULL,
  doTitleBox = TRUE,
  colorSub = NULL,
  titleColor = "black",
  titleFont = 2,
  titlePreset = "top",
  titleAdjPreset = "top",
  xlab = "",
  xlabline = 2,
  ylab = "",
  ylabline = 1.5,
  groupSuffix = NULL,
  highlightPch = 21,
  highlightCex = 1.5,
  highlightColor = "#00AAAA66",
  doHighlightPolygon = FALSE,
  highlightPolygonAlpha = 0.3,
  doHighlightLegend = TRUE,
  smoothPtCol = "#00000055",
  margins = c(2.5, 0.5, 2, 0.2),
  outer_margins = c(0, 2.5, 0, 0.2),
  useRaster = TRUE,
  ncol = NULL,
  nrow = NULL,
  doPar = TRUE,
  las = 2,
  groupedX = TRUE,
  customFunc = NULL,
  filterNA = TRUE,
  filterNAreplacement = NA,
  filterNeg = FALSE,
  noise_floor = 0,
  noise_floor_value = NA,
  filterFloor = NULL,
  filterFloorReplacement = NULL,
  transFactor = 0.18,
  nrpoints = 0,
  smoothScatterFunc = jamba::plotSmoothScatter,
  applyRangeCeiling = TRUE,
  doTxtplot = FALSE,
  ablineV = 0,
  ablineH = c(-2, 0, 2),
  blankPlotPos = NULL,
  fillBackground = TRUE,
  useRank = FALSE,
  apply_transform_limit = 40,
  panel_hook_function = NULL,
  doPlot = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{\code{numeric} object usually a \code{matrix} that contains
values with measurement rows, and sample/observation columns.
For example, with gene or protein expression data, the genes
or proteins (or the assays of genes or proteins) are
represented in rows, and obtained samples are represented
in columns. Alternatively \code{x} can be \code{SummarizedExperiment}
object, used alongside argument \code{assay_name}.}

\item{detail_factor}{\code{numeric} used to adjust the level
of detail, as a multiplier for \code{nbin_factor} and \code{bw_factor}.}

\item{nbin_factor}{\code{numeric} value used to adjust the number of bins
used to display the MA-plots, where values higher than \code{1} increase
the resolution and level of detail, and values below \code{1} decrease
the resolution. Note the number of bins are already adjusted based
upon the square root of the number of plot panels, and \code{nbin_factor}
applied to that value.}

\item{bw_factor}{\code{numeric} used to adjust the resolution of the
2-dimensional bandwidth calculation, where higher values create
more detailed density, and lower values create a smoother density
across the range of data.
In some cases, the \code{ggplot} panel aspect ratio diverges from 1:1,
in which case \code{bw_factor} can be used to expand the bandwidth by
y-axis, or y-axis, respectively. For example, if the density
appears short-wide, try \code{bw_factor=c(1.2, 1)}, if the density
appears tall-skinny, try \code{bw_factor=c(1, 1.2)}.}

\item{assay_name}{\code{character} used when \code{x} is a
\code{SummarizedExperiment} object, to determine which assay
\code{matrix} to use for the MA plots. When \code{assay_name=NULL}
the first assay entry is used, for example \code{assays(x)[[1]]}.}

\item{useMedian}{\code{logical} indicates whether to center data
using the \code{median} value, where \code{useMedian=FALSE} by default.
The median is preferred in cases where outliers should not influence
centering.
The mean is preferred in cases where the data should visualize
data in a manner consistent with downstream parametric
statistical analysis.
When a particular sample represents a technical outlier,
one option to visualize data without being skewed by the outlier
is to define \code{controlSamples} to exclude the outlier sample(s).
In this way, data centering will be applied using the non-outlier
samples as reference.}

\item{controlSamples}{\code{character} vector of \code{colnames(x)} passed
to \code{centerGeneData()} which defines the control samples during
the data centering step.
By default, and the most common practice, MA-plots are
calculated across all samples, which effectively uses all
\code{colnames(x)} as \code{controlSamples}.
However, it is quite useful sometimes to provide a subset of
samples especially if there are known quality samples, to which
new samples of unknown quality are being compared.}

\item{centerGroups}{\code{character} vector of groups passed to
\code{jamma::centerGeneData()} which determines how data is centered.
Each group is centered independently, to enable visual
comparisons within each relevant centering group.
It is useful to center within batches or within
subsets of samples that are not intended to be compared to
one another.
Another useful alternative is to center by each sample
group in order to view the variability among group
replicates, which should be much lower than variability across
sample groups. See \code{centerGeneData()} for more specific examples.}

\item{colramp}{\code{character} input recognized by
\code{jamba::getColorRamp()}. Some examples:
\itemize{
\item name of color palette from RColorBrewer
\item the name of a function from the 'viridis' or 'viridisLite' package,
either as \code{viridisLite::viridis()} or "viridis"
\item a single \code{character} colors, from which a gradient is created
from white (default is \code{defaultBaseColor="white"}) to the
specified color
\item vector of R colors
}}

\item{groupedX}{\code{logical} indicating whether the x-axis value, which
represents the median or mean value, should be calculated independently
for each group when \code{centerGroups} is used with multiple groups.
Typically \code{groupedX=TRUE} is recommended, however it can be beneficial
to share an overall x-axis value in specific circumstances.}

\item{grouped_mad}{\code{logical} indicating whether the MAD factor calculation
of variability among samples should be performed independently
for each group when \code{centerGroups} is used with multiple groups.
Typically \code{grouped_max=TRUE} is recommended, however it can be beneficial
to share an overall MAD factor threshold across all samples
in specific circumstances.}

\item{outlierMAD}{\code{numeric} threshold above which a MA-plot panel
MAD factor is considered an outlier. When a MA-plot panel is
considered an outlier, the \code{outlierColramp} or \code{outlierColor}
is applied to the panel color ramp to display a visual
indication.}

\item{mad_row_min}{\code{numeric} value indicating the minimum x-axis
value, calculated using either median or mean as defined by
argument \code{useMedian}, at or above which a measurement is used in the
MAD factor calculation. This threshold is useful to restrict the
MAD variability calculation to measurements (rows in \code{x}) with
signal that meets a minimum noise threshold.}

\item{displayMAD}{\code{logical} indicating whether to display each
MA-plot panel MAD factor (median absolute deviation). A MAD
value for each panel is calculated by taking the median absolute
deviation from zero across all points, using points whose mean
value is equal or greater than \code{outlierRowMin}. The overall MAD
is defined by the median MAD from the MA-plot panels. The MAD
factor is defined as the ratio of each MA-plot panel MAD value
to the overall MAD value, and therefore most MAD factor values
should be roughly \code{1}. The overall MAD value is defined by the
median across all samples when \code{groupedMAD=FALSE}, or defined
within each \code{centerGroup} when \code{groupedMAD=TRUE}. A value
with MAD factor 2 is interpreted as a sample whose median
deviation from zero is twice as high as the typical sample,
which is a reasonably indication that this sample has twice
the inherent level of noise compared to other samples. Note
that MAD values should be interpreted within sample processing
batches if relevant, or within logical experimental units --
roughly interpreted to mean sets of samples within which direct
statistical comparisons are intended to be applied. For example,
gene expression data that include brain and liver samples would
probably use \code{centerGroups} for brain and liver to be centered
separately, therefore the MAD factors should be separately
calculated for brain and for liver.}

\item{noise_floor, noise_floor_value}{\code{numeric} to define a numeric
floor, or \code{NULL} for no numeric floor. Values \strong{at or below}
\code{noise_floor} are set to \code{noise_floor_value}, intended for two
potential uses:
\enumerate{
\item Filter out value below a threshold, so they do not affect centering.
\itemize{
\item This option is valuable to remove zeros when a zero \code{0} is considered
"no measurement observed", typically for count data such as RNA-seq,
NanoString, and especially single-cell protocols or other protocols
that produce a large number of missing values.
\item One can typically tell whether input data includes zero \code{0}
values by the presence of characteristic 45-degree angle lines
originating from \code{x=0} angled toward the right. The points along
this line are rows with more measurements of zero than non-zero,
there this sample has a non-zero value.
}
\item Set values at a noise floor to the noise floor, to retain the
measurement but minimize the effect during centering to the lowest
realiable measurement for the platform technology.
\itemize{
\item This value may be set to a platform noise floor
for something like microarray data where the intensity may be
unreliable below a threshold; or
\item for quantitative PCR measurements where cycle threshold (Ct)
values may become unreliable, for example above CT=40 or CT=35.
Data is often transformed to abundance with \code{2 ^ (40 - CT)} then
log2-transformed for analysis. In this case, to apply a \code{noise_floor}
effective for CT=35, one would use \code{noise_floor=5}.
}
}}

\item{naValue}{\code{character} string used to convert values of \code{NA} to
something else. This argument is useful when a numeric matrix may
contain \code{NA} values but would prefer them to be, for example, \code{0}.}

\item{centerFunc}{\code{function} used to supply a custom data centering
function. In practice this argument should rarely be changed.}

\item{whichSamples}{\code{NULL} or \code{integer} vector, representing
an index subset of samples to include in the MA-plots. When
\code{whichSamples} represents a subset of samples in \code{x}, the
MA-plot calculations are performed on all samples, then only
samples in \code{whichSamples} are displayed. This argument keeps
the MA-plot calculations consistent even when viewing only
one or a subset of samples in more detail.}

\item{useRank}{\code{logical} indicating whether to create column-wide
ranks, then create MA-plots using the rank data. When \code{useRank=TRUE}
the y-axis represents the rank difference from mean, and the
x-axis represents the mean rank. Using \code{useRank=TRUE} is a good
method to evaluate whether data can be normalized, or whether
data across samples is inherently noisy.}

\item{titleBoxColor}{\code{character} vector of R colors, where
\code{titleBoxColor} is equal to \code{ncol(x)}, or where
\code{names(titleBoxColor)} matches \code{colnames(x)}. When supplied, each
plot panel strip background will be colored accordingly.}

\item{outlierColor}{\code{character} string representing one R color,
used when \code{colrampOutlier} is \code{NULL} and when \code{outlierMAD} is
defined. This color is used for MA-plot outlier panels by
substituting the first color from the \code{colramp} color ramp,
to act as a visual cue that the panel represents an outlier.}

\item{fillBackground}{\code{logical} currently used for base R graphics
output, and passed to \code{jamba::plotSmoothScatter()},
indicating whether to fill the plot panel using the
first color in the color ramp for each MA-plot panel, or when
a plot panel is an outlier, it uses \code{outlierColor}.
This argument is useful for all plot panels especially when the
\code{colramp} base color is not white (or otherwise does not match
the background of the plot device, for example \code{colramp="viridis"}).
This argument is also used with \code{outlierColor} to add visual emphasis
to plot panels where the MAD factor exceeds \code{outlierMAD}.}

\item{maintitle}{\code{character} string with the title displayed above
all individual MA-plot panels. It will appear in the top outer
margin.}

\item{subtitle}{\code{NULL} or \code{character} vector to be drawn at
the bottom left corner of each plot panel, the location
is defined by \code{subtitlePreset}.}

\item{transFactor}{\code{numeric} adjustment to the visual density of
smooth scatter points. For base R graphics, this argument is
passed to \code{jamba::plotSmoothScatter()}. The argument value is based upon
\code{graphics::smoothScatter()} argument \code{transformation}, which uses
default \code{function(x)x^0.25}. The \code{transFactor} is equivalent to the
exponential in the form: \code{function(x)x^transFactor}. Lower values
make the point density more visually intense, higher values make the
point density less visually intense.}

\item{doPlot}{\code{logical} indicating whether to create plots. When
\code{doPlot=FALSE} only the MA-plot panel data is returned.}

\item{highlightPoints}{\code{character} rownames, default NULL, with optional
rows to highlight in each MA-plot panel. Input also recognizes:
\itemize{
\item \code{character} vector matching \code{rownames(x)}, indicating a single
set of points to highlight in one color defined in \code{highlightColor}.
Internally, it is converted to a \code{list} containing one vector.
\item \code{list} of \code{character} vectors, each containing \code{rownames(x)}.
Each vector is highlighted as a set, using colors defined in
\code{highlightColor} which ideally should have length equal to
\code{length(highlightPoints)}.
\item \code{NULL} no points are highlighted.
}}

\item{highlightCex}{\code{numeric} vector used when \code{highlightPoints}
is defined. It is recycled to \code{length(highlightPoints)} after
\code{highlightPoints} is converted to a \code{list} if necessary.
Values are therefore applied to each set of points in the \code{list}.
It can be supplied as a \code{list}, which will be recycled to
\code{length(highlightPoints)} and applied in order to each \code{vector}
of points highlighted.}

\item{highlightColor}{\code{character} vector used when \code{highlightPoints}
is defined. It is recycled to \code{length(highlightPoints)} after
\code{highlightPoints} is converted to a \code{list} if necessary.
Colors are therefore applied to each set of points in the \code{list}.}

\item{doHighlightLegend}{\code{logical} indicating whether to print a
color legend when \code{highlightPoints} is defined. The legend is
displayed in the bottom outer margin of the page using
\code{outer_legend()}, and the page is adjusted to add bottom
outer margin. When the legend is particularly large, it may
be preferable to hide the color legend and display the legend
another way, for example \code{legend()}.}

\item{ablineH, ablineV}{\code{numeric} vector indicating position of
horizontal and vertical lines in each MA-plot panel.
Either argument can be supplied as a \code{list}, which will be applied
to each plot panel in order, which allows specifying specific
abline values in individual panels.}

\item{xlim}{\code{numeric} or \code{NULL} to define a fixed numeric range for
x-axis values. When \code{xlim=NULL} the ranges are defined by the
numeric x-axis values in each \code{centerGroups} grouping, so that
each group can have its own independent x-axis ranges.
When there is a large proportion of values with x=0 (or near zero),
there are two options to reduce the point density near zero,
which can improve visibility of non-zero points:
\enumerate{
\item Use \code{noise_floor=0} and \code{noise_floor_value=NA}, which replaces
values at or below zero with \code{NA}, thereby hiding these points
from each plot panel. Note the \code{NA} values are also not used during
data centering calculations.
\item Define \code{xlim=c(0.001, 20)} and \code{applyRangeCeiling=FALSE}, which
defines the x-axis minimum slightly above zero, and
\code{applyRangeCeiling=FALSE} does not display points outside the
x-axis range at the plot limits, thereby hiding those points.
This method does not replace values with \code{NA}, therefore all
non-NA values are used during data centering.
}}

\item{ylim, xlim}{\code{NULL} or \code{numeric} vector \code{length=2} indicating
the y-axis and x-axis ranges, respectively. The values are useful
to define consistent dimensions across all panels. The
default \code{ylim=c(-4, 4)} represents 16-fold up and down range in
normal space, assuming the data content has already been log2
transformed, and is typically a reasonable starting point for
most purposes. Even if numeric values are all between
\code{-1.5} and \code{1.5}, it is still recommended to keep a range in
context of \code{c(-4, 4)}, or an appropriate and consistent range given
the data content, so the data has some visual context.
The range \code{c(-4, 4)} should be adjusted
relative to the typical ranges expected for the data.}

\item{ncol, nrow}{\code{integer} number of MA-plot panel columns and rows
passed to \code{graphics::par("mfrow")} when \code{doPar=TRUE}. When only one
value is supplied, \code{nrow} or \code{ncol}, the other value is defined
by \code{ncol(x)} and \code{blankPlotPos} so all panels can be contained on
one page. When \code{nrow} and \code{ncol} are defined such that multiple
pages are produced, each page will be annotated with \code{maintitle}
and \code{doHighlightLegend} if relevant.}

\item{blankPlotPos}{\code{NULL} or \code{integer} vector indicating
plot panel positions to be drawn blank, and therefore skipped.
This argument is intended to allow manual placement of plot panels
with spacing that reinforces a logical layout.
Plot panels are drawn in the exact order of \code{colnames(x)} received,
with panels placed into a number of columns and rows defined
with \code{ncol}, \code{nrow} or \code{jamba::decideMfrow()} when \code{doPar=TRUE}.
Blank panel positions are intended to help customize the visual
alignment of MA-plot panels. The mechanism is similar to
\code{ggplot2::facet_wrap()} except that blank positions can be manually
defined by what makes sense for the experiment design.
This method is not particularly user-friendly, but allows fine
control over plot panel spacing.}

\item{verbose}{logical indicating whether to print verbose output.}

\item{...}{additional parameters sent to downstream functions,
\code{jamba::plotSmoothScatter()}, \code{jammacalc()}, \code{centerGeneData()}.}

\item{useMean}{(deprecated) \code{logical}, use \code{useMedian}. This argument
indicates whether to center data using the \code{mean} value.
When \code{useMean=NULL} the argument \code{useMedian} is preferred.
For backward compatibility, when \code{useMean} is not \code{NULL},
then \code{useMedian} is defined by \code{useMedian <- !useMean}.}

\item{outlierRowMin}{\code{numeric} value indicating the minimum mean
value as displayed on the MA-plot panel x-axis, in order for
the row to be included in MAD calculations. This argument is
intended to prevent measurements whose mean value is below
a noise threshold from being included, therefore only including
points whose mean measurement is above noise and represents
"typical" variability.}

\item{groupedMAD}{\code{logical} indicating how the MAD calculation
should be performed: \code{groupedMAD=TRUE} (default) calculates the
median absolute deviation (MAD) from y=0 for each sample per \code{centerGroups}
value, then the median MAD value for all samples in each group.
The corresponding MAD factor is defined within each group.
When \code{groupedMAD=FALSE}, the median MAD value is calculated
across all samples, regardless of \code{centerGroups} value,
from which the MAD factor is defined.}

\item{colrampOutlier}{one of several inputs recognized by
\code{jamba::getColorRamp()} to define a specific color ramp for
MA-plot outlier panels, used when \code{outlierMAD} is defined.
When \code{colrampOutlier} is \code{NULL} the \code{outlierColor} is used.}

\item{maintitleCex}{\code{numeric} cex character expansion used to
resize the \code{maintitle}.}

\item{subtitlePreset}{\code{character} string passed to \code{jamba::coordPresets()}.
The default \code{subtitlePreset="bottomleft"} defines the bottom-left
corner of each panel.}

\item{subtitleAdjPreset}{\code{character} string passed to \code{jamba::coordPresets()}.
The default \code{subtitleAdjPreset="topright"} places labels to the
top-right of the subtitle position, which by default is the bottom-left
corner of each panel.}

\item{doTitleBox}{\code{logical} indicating whether to draw plot
titles using a colored box. When \code{doTitleBox=TRUE} the
\code{jamba::drawLabels()} is called to display a label box at
the top of each plot panel, with \code{drawBox=TRUE}. When
\code{doTitleBox=FALSE}, \code{jamba::drawLabels()} is called with
\code{drawBox=FALSE}.}

\item{colorSub}{\code{character} vector of colors, whose names assign each
color to a character value. Alternatively a \code{list} from which
named colors will be extracted.}

\item{titleColor}{\code{character} vector of colors applied to the title
box in each plot panel, but only when multiple \code{titleColor}
are supplied. Otherwise, the title color is defined by
\code{jamba::setTextContrastColor()} along with \code{titleBoxColor}.}

\item{titleFont}{\code{integer} font compatible with \code{par("font")}.
Values are recycled across panels, so each panel can use a custom
value if needed.}

\item{titlePreset}{\code{character} string passed to \code{jamba::coordPresets()}.
The default \code{titlePreset="top"} defines the top edge of each panel,
and default \code{titleAdjPreset="top"} places the title at the top
edge of this position.}

\item{titleAdjPreset}{\code{character} string passed to \code{jamba::coordPresets()}.
The default \code{titleAdjPreset="top"} places labels above the \code{titlePreset}
location, by default above the top edge of each panel.
To place the title inside (below) the top edge of the plot panel,
use \verb{titlePreset="top", titleAdjPreset="bottom"}, however the
title box may overlap and hide data points in the plot panel.}

\item{xlab, ylab}{\code{character} x- and y-axis labels, respectively.
The default values are blank \code{""} because there are a wide variety
of possible labels, and the labels take up more space
than is often useful for most MA-plots.}

\item{xlabline, ylabline}{\code{numeric} number indicating the text line
distance from the edge of plot border to place \code{xlab} and \code{ylab}
text, as used by \code{graphics::title()}.}

\item{groupSuffix}{(deprecated) \code{character} text appended to each MA-plot
panel title. Use argment \code{subtitle} as the preferred alternative.}

\item{doHighlightPolygon}{\code{logical} indicating whether to draw
a shaded polygon encompassing \code{highlightPoints}, using each
\code{highlightColor}.The polygon is defined by \code{grDevices::chull()}
via the function \code{points2polygonHull()}.}

\item{highlightPolygonAlpha}{\code{numeric} value indicating alpha
transparency used for the highlight polygon when
\code{doHighlightPolygon=TRUE}, where \code{0} is fully transparent,
and \code{1} is completely not transparent (opaque).}

\item{smoothPtCol}{\code{color} used to draw points when \code{nrpoints} is
non-zero, which draws points in the extremities of the
smooth scatter plot. See \code{jamba::plotSmoothScatter()}.
The effect can also be achieved by adjusting \code{transFactor} to
a lower value, which increases the visual contrast of individual
points in the point density.}

\item{margins}{\code{numeric} vector of margins compatible with
\code{graphics::par("mar")}.
Default values are provided here for convenience.
Since version \verb{0.0.31.900} the \code{outer_margins} are used to control
y-axis label whitespace, and and y-axis labels are not displayed
on every plot panel when \code{doPar=TRUE}.}

\item{outer_margins}{\code{numeric} vector of margins compatible with
\code{graphics::par("oma")}.
Default values are provided here for convenience.
The left outer marging is used to allow whitespace to display
y-axis labels.
Note when \code{useRank=TRUE} the y-axis whitespace is increased
to accomodate larger integer label values.}

\item{useRaster}{\code{logical} indicating whether to draw the
smooth scatter plot using raster logic, \code{useRaster=TRUE} is
passed to \code{jamba::plotSmoothScatter()}. The default \code{TRUE}
creates a much smaller plot object by rendering each plot
panel as a single raster image instead of rendering individual
colored rectangles. There is no driving reason to use \code{useRaster=FALSE}
except if the rasterization process itself is problematic.}

\item{doPar}{\code{logical} indicating whether to apply
\code{graphics::par("mfrow")} to define MA-plot panel rows and columns.
When \code{doPar=FALSE} each plot panel is rendered without
adjusting the \code{graphics::par("mfrow")} setting, which is appropriate
for displaying each plot panel individually.}

\item{las}{\code{integer} value \code{1} or \code{2} indicating whether axis labels
should be parallel or perpendicular to the axes, respectively.}

\item{customFunc}{optional \code{function} used instead of the summary
function defined by \code{useMedian} during the data centering step.
This function should take a \code{matrix} of \code{numeric} values as input,
and must return a \code{numeric} vector length equal to \code{nrow(x)}.
This option is intended to allow custom row statistics,
for example geometric mean, or other row summary functions.}

\item{filterNA, filterNAreplacement}{\code{logical} and \code{vector} respectively.
When \code{filterNA=TRUE}, all \code{NA} values are replaced with
\code{filterNAreplacement}.
This process is conceptually opposite of \code{noise_floor} which replaces
a \code{numeric} value with \code{NA} or another \code{numeric} value.
Instead, \code{filterNA} is intended to convert \code{NA} values into
a known \code{numeric} value, typically using an appropriate noise floor
value such as zero \code{0}.
Practically speaking, \code{NA} values should probably
be left as \code{NA} values, so that data centering does not use these values,
and so the MA-plot panel does not draw a point when no measurement
exists.}

\item{filterNeg}{(deprecated) \code{logical} argument, use \code{noise_floor}.}

\item{filterFloor, filterFloorReplacement}{(deprecated) in favor of
\code{noise_floor}, and \code{noise_floor_replacement} respectively.}

\item{nrpoints}{\code{integer} or \code{NULL} indicating the number of points
to display on the extremity of the smooth scatter density,
passed to \code{jamba::plotSmoothScatter()}.}

\item{smoothScatterFunc}{\code{function} used to produce a smooth scatter plot
in base R graphics. The default \code{jamba::plotSmoothScatter()} controls
the level of detail in the density calculation, and in the graphical
resolution of that density in each plot panel. The custom function
should accept argument \code{transformation} as described in \code{transFactor},
even if the argument is not used. This argument could be useful to
specify customizations not convenient to apply otherwise.}

\item{applyRangeCeiling}{\code{logical} passed to
\code{jamba::plotSmoothScatter()} which determines how to handle points
outside the plot x-axis and y-axis range: \code{applyRangeCeiling=TRUE}
will place points at the border of the plot, which is helpful
to indicate that there are more points outside the viewing range;
\code{applyRangeCeiling=FALSE} will crop and remove points outside
the viewing range, which is helpful for example when a large
number of points are at zero and overwhelm the point density.
When there are a large proportion of values at zero, it
can be helpful to apply \code{xlim=c(0.01, 20)} and
\code{applyRangeCeiling=FALSE}.}

\item{doTxtplot}{\code{logical} (not yet implemented in \code{jamma}),
indicating to produce colored ANSI text plot output, for example
to a text terminal.}

\item{panel_hook_function}{optional custom \code{function} called
as a "hook" after each MA-plot panel has been drawn. This function
can be used to display custom axis labels, add plot panel visual
accents, lines, labels, etc.
This \code{panel_hook_function} is recycled as a \code{list} to the
number of samples \code{ncol(x)}, which can be used to supply a unique
function for each panel. Any element in the \code{list} with \code{length=0}
or \code{NA} is skipped for the corresponding MA-plot panel.
This function should accept at least two arguments, even if ignored:
\itemize{
\item \code{i} - an \code{integer} indicating the sample to be plotted in order,
as defined by \code{colnames(x)}, and \code{whichSamples} in the event samples
are subsetted or re-ordered with \code{whichSamples}.
\item \code{...} additional arguments passed by \code{...} into this custom function.
\item Any arguments of \code{jammaplot()} are available inside the panel
hook function as a by-product of calling this function within
the environment of the active \code{jammaplot()}, therefore any
argument values will be available for use inside that function.
}}

\item{titleBoxColor, subtitleBoxColor}{\code{character} vector of
R colors used as background color for each panel title text,
or subtitle text respectively. The subtitle appears in the
bottom-left corner, and usually indicates the center groups
as defined by \code{centerGroups}.}
}
\value{
\code{list} of \code{numeric} \code{matrix} objects, one for each MA-plot,
with colnames \code{"x"} and \code{"y"}. This \code{list} is sufficient input
to \code{jammaplot()} to re-create the full set of MA-plots.
}
\description{
Produce MA-plot of omics data, where \code{jammaplot()} uses base R graphics,
\code{ggjammaplot()} uses ggplot2 graphics.
}
\details{
\code{jammaplot} takes a numeric matrix, typically of gene expression data,
and produces an MA-plot (Bland-Altman plot), also known as a
median-difference plot. One panel is created for each column of
data. Within each panel, the x-axis represents the mean or median
expression of each row; the y-axis represents the difference from
mean or median for that column.

By default, the plot uses \code{jamba::plotSmoothScatter()}, with optional
highlighted points draw using \code{points()}.

The function will determine an appropriate layout of plot panels,
which can be overridden using \code{ncol} and \code{nrow} to specify the
number of columns and rows of plot panels, respectively. For now,
this function uses base R graphics instead of ggplot2, in order
to accomodate some custom features.

This function uses "useRaster=TRUE" by default, which causes
\code{jamba::plotSmoothScatter()} to render a rasterized image as opposed
to a composite of colored rectangles. This process substantially
reduces the render time in all cases, and reduces the image size
when saving as PDF or SVG.
\subsection{Notable features}{
\subsection{Highlighting points}{

Specific points can be highlighted with argument \code{highlightPoints}
which can be a vector or named list of vectors, containing \code{rownames(x)}.
When using a list, point colors are assigned to each element in the
list in order, using the argument \code{highlightColor}.
}

\subsection{Centering by control samples}{

Typical MA-plots are "global-centered", which calculates the
mean/median across all columns in \code{x}, and this value is subtracted
from each individual value per row.

By specifying \code{controlSamples}
the mean/median is calculated using only the \code{colnames(x)} which match
\code{controlSamples}, thus representing "difference from control."

It may also be useful to center data by known high-quality samples,
so the effect of potential outlier samples is avoided.
}

\subsection{Centering within subgroups}{

By specifying \code{centerGroups} as a vector of group names,
the centering is calculated within each group of \code{colnames(x)}.
In this way, subsets of samples can be treated independently in
the MA-plots. A good example might be producing MA-plots for
"kidney" samples, and "muscle" samples, which may have
fundamentally different signal distributions. A good rule
of thumb is to apply \code{centerGroups} to represent separate
groups of samples where you do not intend to apply direct
statistical comparisons across those samples, without at
least applying a two-way contrast, a fold change of fold
changes.

Another informative technique is to center by sample group,
for example \code{centerGroups=sample_group}.
This technique produces MA-plots that depict the
"difference from group" for each sample replicate of a sample
group, and is very useful for identifying sample replicates
with markedly higher variability to its sample group than
others. In general, the variability within sample group
should be substantially lower than variability across
sample groups. Use \code{displayMAD=TRUE} and \code{outlierMAD=2}
as a recommended starting point for this technique.
}

\subsection{Applying a noise floor}{

The argument \code{noise_floor} provides a numeric lower threshold,
where individual values \strong{at or below} this threshold are
set to a defined value, defined by argument \code{noise_floor_value}.
The default was updated in version \verb{0.0.21.900} to
\code{noise_floor=0} and \code{noise_floor_value=NA}.
Values of zero \code{0} are set to \code{NA} and therefore are not included
in the MA-plot calculations. Only points above zero are included
as points in each MA-plot panel.

Another useful alternative is to define \code{noise_floor_value=noise_floor}
which sets any measurement \strong{at or below} the \code{noise_floor} to
this value. This option has the effect of reducing random noise from
points that are already below the noise threshold and therefore
are unreliable for this purpose.
}

\subsection{Customizing the panel layout}{

Panels are drawn using the order of \code{colnames(x)} by row,
from left-to-right, then top-to-bottom.
The argument \code{blankPlotPos} is intended to insert an empty panel
at a particular panel position, to help customize the alignment
of sample panels.
This option is typically used with \code{ncol} and \code{nrow} to define
a fixed layout of panel columns and rows. \code{blankPlotPos} refers
to panels numbered as drawn per row of panels,
}

\subsection{Identifying potential sample outliers}{

Use argument \code{displayMAD=TRUE} to display the per-sample MAD factor
relative to its \code{centerGroups} value, if provided. The MAD value
for each MA-plot panel is calculated using rows whose mean
is at or above \code{outlierRowMin}. The median MAD value is calculated
for each \code{centerGroups} grouping when \code{groupedMAD=TRUE}, by default.
Finally, each MA-plot panel MAD factor is the ratio of its MAD value
to the relevant median MAD value. MA-plot panels with MAD factor
above \code{outlierMAD} are considered outliers, and the color ramp
uses \code{outlierColramp} or \code{outlierColor} as a visual cue.

Putative outlier samples should typically \strong{not} be determined when:
\itemize{
\item Data are not yet normalized. One major purpose of MA-plots is to
assess relative signal (divergence from y=0 across sample panels).
The outlier MAD calculation is based upon the assumption that each
panel is already normalized as best as possible to y=0.
\item \code{controlSamples} is defined to include only a subset
of sample groups,
\item \code{centerGroups} is not defined, or represents more than one
set of sample groups that are not intended to be statistically
compared directly to one another.
}

Putative outlier samples may be defined when:
\itemize{
\item Data are appropriately normalized.
\item There are sufficient measurements that it is reasonable to assume
that the mean or median signal is centered at y=0.
\item \code{centerGroups} represents sets of sample groups that are
intended to be involved in direct comparisons.
\item \code{centerGroups} represents each sample group.
}

Potential sample outliers may be identified by setting a threshold
with \code{outlierMAD}, by default \code{outlierMAD=5}. For a sample to be considered
an outlier, its median difference from mean/median needs to be
\strong{5} times higher than the median across samples.

That said, we typically recommend an \code{outlierMAD=2} when centering
by sample groups, or when centering within experimental subsets.
For one sample to have 2xMAD factor, its variance needs
to be twice as high as the majority of other samples, which
is symptomatic of potential technical failure.

There are exceptions to this suggested guideline, which includes
scenarios where a batch effect may be involved, use of clinical or
animal samples with relatively small sample sizes, or scenarios
in which variability is known or strongly suspected to be biological
variance and not technical variance.
\itemize{
\item Biological variability is to be maintained.
\item Technical variability is also to be maintained, but only where the
technical variability is reasonably consistent from sample to sample.
}

To do:
\itemize{
\item{Accept other object types as input, including Bioconductor
classes: \code{ExpressionSet} (DONE), \code{SummarizedExperiment} (DONE),
\code{MultiExperimentSet} (TODO)}
\item{Make it efficient to convey group information, for example
define \code{titleBoxColor} with group colors, allow \code{centerByGroup=TRUE}
which would re-use known sample group information.}
\item{Adjust the suffix to indicate when \code{centerGroups} are being
used. For example indicate \code{'sampleID vs groupA'} instead of
\code{'sampleID vs median'}.}
}
}

}
}
\section{Functions}{
\itemize{
\item \code{ggjammaplot()}: 

}}
\examples{
if (requireNamespace("SummarizedExperiment", quietly=TRUE)) &&
   requireNamespace("farrisdata", quietly=TRUE)) {
   suppressPackageStartupMessages(require(SummarizedExperiment));

   GeneSE <- farrisdata::farrisGeneSE;

   titleBoxColor <- jamba::nameVector(
      farrisdata::colorSub[as.character(
         SummarizedExperiment::colData(GeneSE)$groupName)],
      colnames(GeneSE));
   options("warn"=FALSE);

   gg <- ggjammaplot(GeneSE,
      ncol=6,
      base_size=12,
      maintitle="Farris raw RNAseq data",
      titleBoxColor=jamba::rgb2col(col2rgb(titleBoxColor)),
      assay_name="raw_counts")

   gg <- ggjammaplot(GeneSE,
      ncol=6,
      assay_name="counts",
      useRank=TRUE,
      ylim=c(-11000, 11000),
      maintitle="MA-plots by rank and rank difference",
      titleBoxColor=titleBoxColor)

   gg <- ggjammaplot(GeneSE,
      ncol=6,
      assay_name="counts",
      titleBoxColor=titleBoxColor,
      base_size=10,
      maintitle="MA-plots showing MAD factor",
      displayMAD=TRUE)

   gg <- ggjammaplot(GeneSE,
      ncol=6,
      assay_name="counts",
      titleBoxColor=titleBoxColor,
      maintitle="MA-plot omitting one panel, then using blankPlotPos",
      whichSamples=colnames(GeneSE)[c(1:21, 23:24)],
      blankPlotPos=22,
      displayMAD=TRUE)

   if (FALSE) {
   ggdf <- ggjammaplot(GeneSE,
      assay_name="counts",
      whichSamples=c(1:3, 7:9),
      return_type="data",
      titleBoxColor=titleBoxColor)
   highlightPoints1 <- names(jamba::tcount(subset(ggdf, mean > 15 & difference < -1)$item, 2))
   highlightPoints2 <- subset(ggdf, name \%in\% "CA1CB492" &
      difference < -4.5)$item;
   highlightPoints <- list(
      divergent=highlightPoints1,
      low_CA1CB492=highlightPoints2);

   ggdf_h <- ggjammaplot(GeneSE,
      assay_name="counts",
      highlightPoints=highlightPoints,
      whichSamples=c(1:3, 7:9),
      return_type="data",
      titleBoxColor=titleBoxColor)

   # you can use output from `jammaplot()` as input to `ggjammaplot()`:
   jp2 <- jammaplot(GeneSE,
      outlierMAD=2,
      doPlot=FALSE,
      assay_name="raw_counts",
      filterFloor=1e-10,
      filterFloorReplacement=NA,
      centerGroups=colData(GeneSE)$Compartment,
      subtitleBoxColor=farrisdata::colorSub[as.character(colData(GeneSE)$Compartment)],
      useRank=FALSE);

   gg1 <- ggjammaplot(jp2,
      ncol=6,
      titleBoxColor=titleBoxColor);
   print(gg1);
   }
}

# Note the example data requires the affydata Bioconductor package
if (suppressPackageStartupMessages(require(affydata))) {
   data(Dilution);
   edata <- log2(1+exprs(Dilution));
   jammaplot(edata);

   jammaplot(edata,
      whichSamples=c(1, 2));

   jammaplot(edata,
      sample_labels=paste("Sample", colnames(edata)));

   jammaplot(edata,
      controlIndicator="titlestar");

   jammaplot(edata,
      controlIndicator="none");

   jammaplot(edata,
      panel_hook_function=function(i,...){box("figure")});

   jammaplot(edata,
      useRank=TRUE,
      maintitle="Rank MA-plots");
}

}
\seealso{
\code{jamba::plotSmoothScatter()}

Other jam plot functions: 
\code{\link{volcano_plot}()}

Other jam plot functions: 
\code{\link{volcano_plot}()}
}
\concept{jam plot functions}
