% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamma.r
\name{jammaplot}
\alias{jammaplot}
\alias{MAplot,}
\alias{MVAplotMED,}
\alias{MVAplotMEDsmooth,}
\alias{MVAplot,}
\alias{plotMA,}
\alias{plotMVA,}
\alias{ma.plot,}
\alias{plot.ma,}
\alias{mva.plot}
\title{Produce MA-plot of omics data.}
\usage{
jammaplot(object, colramp = c("white", "lightblue", "blue", "navy",
  "orange", "orangered2"), colrampOutlier = NULL,
  outlierColor = "palegoldenrod", whichSamples = NULL,
  maintitle = NULL, subtitle = NULL, subtitlePreset = "bottom",
  maintitleSep = "\\n", titleCexFactor = 1, titleCex = NULL,
  doTitleBox = TRUE, titleBoxColor = "#DDBB9977",
  titleColor = "black", titleFont = 2, titlePreset = "top",
  xlab = "", xlabline = 2, ylab = "", ylabline = 1.5,
  groupSuffix = " vs. Median", highlightPoints = NULL,
  highlightPch = 21, highlightCex = 1, highlightColor = "#00AAAA66",
  doHighlightPolygon = FALSE, highlightPolygonAlpha = 0.3,
  smoothPtCol = "#00000055", margins = c(3.5, 3, 0.7, 0.5),
  useRaster = TRUE, ncol = NULL, nrow = NULL, doPar = TRUE,
  las = 2, ylim = c(-4, 4), xlim = NULL,
  controlSamples = colnames(object), centerGroups = NULL,
  groupedX = TRUE, useMean = FALSE, customFunc = NULL,
  filterNA = TRUE, filterNAreplacement = NA, filterNeg = TRUE,
  filterFloor = 0, filterFloorReplacement = filterFloor,
  transFactor = 0.18, nrpoints = 50,
  smoothScatterFunc = jamba::plotSmoothScatter,
  applyRangeCeiling = TRUE, doTxtplot = FALSE, ablineV = 0,
  ablineH = c(-2, 0, 2), blankPlotPos = NULL, outlierMAD = 5,
  outlierRowMin = 5, groupedMAD = TRUE, displayMAD = FALSE,
  fillBackground = TRUE, doPlot = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{object}{numeric matrix typically containing log-normal measurements,
with measurement rows, and sample columns.}

\item{colramp}{one of several inputs recognized by
\code{\link{getColorRamp}}. It typically recognizes either the name of
a color ramp from RColorBrewer, the name of functions from the
\code{\link[viridis]{viridis}} package, or single R colors, or
a vector of R colors.}

\item{colrampOutlier}{one of several inputs recognized by
\code{\link{getColorRamp}}, to be used only in panels determined to be
outliers based upon MAD outlier logic. Typically the color ramp is
identical to \code{colramp} except the background color indicates
an error, for example yellow background instead of white.}

\item{outlierColor}{character R color, used when \code{colrampOutlier is NULL}
to substitute as the first color from \code{colramp}. This method keeps the
color ramp consistent, but changes the background color to highlight
the plot panel as an outlier.}

\item{whichSamples}{NULL or integer vector, representing an index of samples
to include in the MA-plots. Note that all samples are used to create
the MA-plot data, but only these samples will be plotted, which is useful
when wanting to view only a subset of samples in detail, but where the
MA-plot data is identical to the full dataset.}

\item{maintitle}{the main title for each panel, either printed inside each
title box when doTitleBox=TRUE, or atop each panel when doTitleBox=FALSE.}

\item{subtitle}{NULL or character text to be drawn at the bottom center
of each plot panel.}

\item{subtitlePreset}{character value describing where to position the
subtitle, using terms valid in \code{jamba::coordPresets()}. The default "bottom"
centers the subtitle at the bottom of each panel.}

\item{maintitleSep}{character separator used to combine multiple values
for the maintitle, used to keep all information either on one line or
split on multiple lines.}

\item{doTitleBox}{logical whether to draw plot titles using a colored box.}

\item{titleBoxColor}{vector of colors applied to title text. When doTitleBox=TRUE
one or no value is supplied, it defines colors using
\code{\link{jamba::setTextContrastColor}} to use a contrasting color.}

\item{titleColor}{vector of colors applied to title text. When doTitleBox=TRUE
one or no value is supplied, it defines colors using
\code{\link{jamba::setTextContrastColor}} to use a contrasting color.}

\item{titleFont}{integer font compatible with \code{par("font")}. Values
are recycled across panels.}

\item{titlePreset}{character value describing where to position the
subtitle, using terms valid in \code{jamba::coordPresets()}. The default is
"top" which centers the label at the top of each panel.}

\item{xlab, ylab}{character x- and y-axis labels.}

\item{xlabline, ylabline}{numeric number of text lines as used by
\code{title} to position labels relative to the plot area.}

\item{groupSuffix}{character text appended to each plot panel title,
useful for indicating how a sample was processed, beyond just using
colnames. For example, \code{groupSuffix=" vs. group median"} could
indicate that a sample is centered relative to its group median, rather
than the global median.}

\item{highlightPoints}{NULL, character vector, or list of character vectors,
containing points to highlight based upon \code{rownames(object)}. If
a list is supplied, then colors from \code{highlightColors} are recycled
to \code{length(highlightPoints)} so that each list element can receive
its own color. This coloring may be useful for coloring several subsets
of rows, for example housekeeper genes, ribosomal genes, or platform
positive- and negative-control genes.}

\item{highlightColor, highlightCex}{values recycled to the length of
highlightPoints. If highlightPoints is a list, then a list or vector
can be supplied.}

\item{doHighlightPolygon}{logical whether to draw a polygon encompassing
highlighted points. If highlightPoints is a list, a polygon is drawn
around each list element subset of points, using the appropriate
highlightColors for each list element. The polygon is defined by
\code{grDevices::chull()}.}

\item{highlightPolygonAlpha}{numeric value between 0 and 1 indicating the
alpha transparency to use for the polygon color, based upon
highlightColor.}

\item{smoothPtCol}{color used when nrpoints>0, and
\code{\link{jamba::plotSmoothScatter}} draws this many points in the extremities.}

\item{margins}{vector of margins compatible with \code{par("mar")}. Defaults
are applied, but provided here for convenient override.}

\item{useRaster}{logical whether to define \code{useRaster=TRUE} in
\code{\link{jamba::plotSmoothScatter}}, which ultimately calls
\code{\link{imageDefault}}. When TRUE, it creates a much smaller
plot object, because it returns a raster image instead of a set of
rectangles when rendering the image.}

\item{ncol, nrow}{integer number of columns and rows used by
\code{par("mfrow")} when \code{doPar=TRUE} is defined. These values
are helpful in defining a fixed layout for multiple MA-plot panels.}

\item{doPar}{logical whether \code{par("mfrow")} should be defined, in
order to fit multiple panels in one plot window. Set to FALSE if the plot
layout is already defined, or if plotting one panel per page, for
example.}

\item{las}{integer value 1 or 2, indicating whether axis labels should be
parallel or perpendicular to the axes, respectively. This default value
is defined and provided here for convenience to override it.}

\item{ylim, xlim}{NULL or vector length 2 indicating the y- and x-axis
ranges, respectively. These values are useful to define upfront when
it is best to focus on fixed, consistent ranges across all samples. The
default \code{ylim=c(-4,4)} represents 16-fold range in normal space, and
is typically a reasonable starting point for most purposes. If values are
all between -1.5 and 1.5, then it is still good to keep that range in
context of -4 and 4, to indicate that the observed noise is lower than
typically observed. It is also helpful to use xlim to trim off zeros
for some data where there might be many undetected entries, combined
with \code{applyRangeCeiling=FALSE} which also removed outlier points
from the plot density calculation. The result can notably boost the
detail displayed in the plot range.}

\item{controlSamples}{vector of colnames(object) which should be used
as control samples in the data centering step. Typical MA-plots are
calculated relative to the overall mean, however it can be insightful
to calculate values relative to a known control group.}

\item{useMean}{logical whether to perform centering using mean or
median. The default median is intended to represent the data without
being skewed by outliers, however it can be informative to use the mean
values when there are few replicates, or when there are no particular
sample outliers.}

\item{customFunc}{an optional function that is used instead of mean or
median. It should take a matrix input, and return a numeric vector
output summarizing each row in \code{object}. It is intended to
provide custom row statistics, for example geometric mean, or other
row summary function.}

\item{filterNA, filterNeg}{logical indicating whether to remove NA or
negative values before creating the MA-plot. It can be useful to
filter negative values, which are often defined as noise in upstream
normalization steps, so they do not impart visible features into the
MA-plot since they are based upon noise. Negative values are set to
zero when filtered, which gives a characteristic 45 degree angle
symmetry above and below zero.}

\item{filterFloor, filterFloorReplacement}{numeric or NULL, when numeric
values in \code{object} are below \code{filterFloor}, they are replaced
with \code{filterFloorReplacement}. This mechanism is an alternative to
\code{filterNeg}, in that it allows defining a filter above or below
zero. For some platform data , it might be useful to define a
\code{filterFloor} roughly equivalent to its noise threshold.}

\item{transFactor}{numeric adjustment used by
\code{\link{jamba::plotSmoothScatter}}. The default is defined, but available
here as a convenient override. The value is used in a power function in
the form \code{transformation=function(x)x^transFactor}. Lower values
make the scatterplot point density more intense.}

\item{nrpoints}{integer or NULL, the number of points to display on the
extremity, as used by \code{\link{jamba::plotSmoothScatter}}.}

\item{smoothScatterFunc}{function used to plot smooth scatter plot, by
default \code{\link{jamba::plotSmoothScatter}}.}

\item{applyRangeCeiling}{see \code{\link{jamba::plotSmoothScatter}} for details,
logical whether to apply a noise floor and ceiling to display points
outside the plot region at the boundaries of the plot. This parameter
is useful to set FALSE when cropping a plot, in order to remove outlier
points from the density calculation. For example when there is a
very large number of zero values, it can be helpful to define
\code{xlim=c(0.1,20), applyRangeCeiling=FALSE}.}

\item{doTxtplot}{logical whether to plot results in colored text output,
currently under development and is still experimental.}

\item{ablineH, ablineV}{numeric vector indicating where to draw indicator
lines, at these horizontal or vertical positions, respectively.}

\item{blankPlotPos}{NULL or integer vector indicating which plot panels
should be used as blank filler positions. These blank panels can be
used to help organize multiple plot panels to indicate sample groups.
A mechanism similar to ggplot2 facets, except with the ability to
insert known blank positions for organization. The facet_grid mechanic
may be used in future, however is not always ideal depending upon the
experiment design, and so for now a simple manual method is provided.}

\item{outlierMAD}{the MAD factor threshold to define a sample an outlier.}

\item{outlierRowMin}{the minimum value as displayed on the x-axis, for
a row in \code{object} to be used in the MAD outlier calculations. This
threshold is intended to define a noise threshold, so that outlier
samples are determined by using rows above the noise threshold. Thus
a sample outlier would need to have greater than five times median
absolute deviation compared to its center group, based only upon
measurements above noise.}

\item{groupedMAD}{logical indicating whether the MAD calculation should
operate only within \code{centerGroups} (\code{groupedMAD=TRUE}) or whether the
MAD calculation is shared across all samples (\code{groupedMAD=FALSE}).}

\item{displayMAD}{logical whether to display MAD (median absolute deviation)
values for each panel. MAD values are defined for each centerGroup if
supplied, or globally if not. The MAD value is CALCULATED by taking the
median deviation from zero as displayed on the y-axis of each MA-plot
panel, then taking the median of those values. Then a MAD factor is
calculated for each MA-plot panel by taking its median devation divided
by the MAD value for its centerGroup. Most samples should therefore be
close to 1. A value of 2 is interpreted as "two times the MAD factor for
its center group." An outlierMAD of 5x is a reasonable cutoff, which
roughly says that sample has 5 times higher median deviation than other
samples in its center group. If all samples are noisy, it would define
a high MAD value, and therefore all samples would be expected to have
similar MAD factors, near 1.}

\item{fillBackground}{logical sent to \code{\link{jamba::plotSmoothScatter}}
indicating whether to fill the plot panel with the background color,
which is typically the first value in the \code{colramp}. Set to TRUE
when axis ranges may be defined which are outside the range of values
for any particular column in \code{object}.}

\item{verbose}{logical indicating whether to print verbose output.}

\item{...}{additional parameters sent to downstream functions,
\code{\link{jamba::plotSmoothScatter}}, \code{\link{centerGeneData}}.}

\item{centerGoups}{vector of groups, of \code{length(colnames(object))}
indicating optional subgroups to use when performing data centering.
For example it may be appropriate to center samples of one cell type
only relative to other samples of the same cell type. See
\code{\link{centerGeneData}} for more specifics.}
}
\value{
List of data used by jamma() to produce an MA-plot, sufficient
to use as input to produce another MA-plot.
}
\description{
Produce MA-plot of omics data.
}
\details{
\code{jammaplot} takes a numeric matrix, typically of gene expression data,
and produces an MA-plot (Bland-Altman plot).

Purpose is to provide an MA plot, based upon the MVAplotMED function from
the now-defunct Agi4x44PreProcess package,  except in this case using
smoothScatter since 20k to 40k datapoints is too many to show without
them being occluded.

You must manually specify the number columns and rows, and set doPar=TRUE
for this function to define multi-panel plotting, e.g.
\code{"nrow=5, ncol=4, doPar=TRUE"}
Otherwise, you can manually set these parameters with
\code{"par('mfrow'=c(5,4));"}

This function also uses "useRaster=TRUE" by default which causes the
\code{\link{imageDefault}} function used by \code{\link{smoothScatter}}
to produce a temporary raster image which is then resized with
interpolation to produce a properly-blended representation of the full
image. This process also substantially reduces the image size when saving
as a vector output format like PDF or SVG.

To highlight certain points, e.g. marked outliers, or statistical hits,
supply rownames into highlightPoints as a vector, or a list of vectors.
If using a list of vectors, also supply a list of colors in
highlightColor, so each subset is colored uniquely.

ctrlSamples is deprecated in favor of controlSamples.

controlSamples indicates which samples should be used for centering
data, by default all samples. However, centering by control samples
visually indicates changes from those control samples.

centerGroups is used to apply the centering to subsets of columns
of data, for example if two very different cell types should be separately
centered for closer scrutiny of the within-cell variability.
One can take this further, and center each sample group, which has
the helpful effect of displaying variability of replicates per
sample group. This method is quite useful for reviewing whether any
replicates are notably different than counterparts in the same
sample group.

Note: filterNeg=TRUE will floor log2 values at 0, which has the
effect of adding the "45 degree lines" originating from c(0,0).

blankPlotPos is intended for specific plot layouts, to keep the plots
organized by some fixed criteria, e.g. certain sample groups per
row or column, but where there may not be consistent replicates per
group.
For ncol and nrow, blankPlotPos indicates which panels should be
left empty, where each panel is numbered by row.  For example,
if ncol=4, then the panels are numbered 1-4, 5-8, 9-12, etc.

outlierMAD is only viable when centerEachGroup=TRUE; in which case
outlierMAD is a threshold for the MAD factor of outliers.  Each MA
plot panel is given a MAD score, using probes whose x-values are
greater than outlierRowMin;, the median of the absolute value is
used as the MAD.
These MAD values are used to calculate the global MAD, and individual
MADs are divided by global MAD to determine a MAD factor.
The outlierMAD is the threshold above which a MAD factor is considered
an outlier.

To do:
\itemize{
\item{Accept other object types as input, including Bioconductor
classes: ExpressionSet, SummarizedExperiment, MultiExperimentSet.}
\item{Make it efficient to convey group information, for example
define titleBoxColor with group colors, allow centerByGroup==TRUE
which would re-use known sample group information.'}
\item{Adjust the suffix to indicate when \code{centerGroups} are being
used. For example indicate 'sampleID vs groupA' instead of
'sampleID vs median'.}
}
}
\examples{
# Note the example data requires the affydata Bioconductor package
if (suppressPackageStartupMessages(require(affydata))) {
   data(Dilution);
   edata <- log2(1+exprs(Dilution));
   jammaplot(edata);
}

}
\seealso{
\code{jamba::plotSmoothScatter()} for enhanced
\code{graphics::smoothScatter()}, \code{centerGeneData()},
\code{jamba::imageDefault()}, \code{jamba::nullPlot()}

Other jam plot functions: \code{\link{points2polygonHull}}
}
\concept{jam plot functions}
