% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-volcano-plot.R
\name{logAxis}
\alias{logAxis}
\title{Log-scaled axis including transformed P-values}
\usage{
logAxis(
  side,
  at = NULL,
  base = 2,
  values = FALSE,
  useFcValues = TRUE,
  las = 2,
  makeNegative = FALSE,
  doSignedSignificance = FALSE,
  flipSignedSignificanceSign = FALSE,
  bigMark = ",",
  prettyN = 5,
  digits = 2,
  cex.axis = 1,
  font.axis = 1,
  ...
)
}
\description{
Log-scaled axis including transformed P-values
}
\seealso{
Other jam utility functions: 
\code{\link{blockArrowMargin}()},
\code{\link{find_colname}()},
\code{\link{fold_to_log2fold}()},
\code{\link{get_se_assaydata}()},
\code{\link{gradient_rect}()},
\code{\link{handle_highlightPoints}()},
\code{\link{log2fold_to_fold}()},
\code{\link{outer_legend}()},
\code{\link{points2polygonHull}()},
\code{\link{update_function_params}()},
\code{\link{update_list_elements}()}
}
\concept{jam utility functions}
