% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamma.r
\name{points2polygonHull}
\alias{points2polygonHull}
\title{define a polygon hull around points}
\usage{
points2polygonHull(x, returnClass = c("matrix", "SpatialPolygons"), ...)
}
\arguments{
\item{x}{two-column numeric matrix of data points, or an object of
class SpatialPoints.}

\item{returnClass}{character class name to return, either matrix or
SpatialPolygons.}

\item{verbose}{logical whether to print verbose output'}
}
\description{
Defines a polygon hull to encompass a set of data points
}
\details{
A polygon hull can be useful as a visual indicator of the location
of a set of points. This function creates a polygon around all points,
and not just the most dense region of points, using the
\code{\link[rgeos]{gConvexHull}} function. Since that function returns
an object of class SpatialPolygons, this function is mainly a wrapper
which returns the set of coordinates, either as a matrix or as the
SpatialPolygons object type.
}
