% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamma.r
\name{points2polygonHull}
\alias{points2polygonHull}
\title{define a polygon hull around points}
\usage{
points2polygonHull(x, returnClass = c("matrix"), ...)
}
\arguments{
\item{x}{two-column numeric matrix of data points, or an object of
class SpatialPoints.}

\item{returnClass}{character class name to return, either matrix or
SpatialPolygons.}

\item{verbose}{logical whether to print verbose output'}
}
\value{
A numeric matrix of polygon coordinates.
}
\description{
Defines a polygon hull to encompass a set of data points
}
\details{
A polygon hull can be useful as a visual indicator of the location
of a set of points. This function creates a polygon around all points,
and not just the most dense region of points, using
\code{rgeos::gConvexHull()}. Since that function returns
an object of class SpatialPolygons, this function is mainly a wrapper
which returns the set of coordinates, either as a matrix or as the
SpatialPolygons object type.
}
\examples{
set.seed(123);
xy <- cbind(x=1:10,
   y=sample(1:10, 10));
xyhull <- points2polygonHull(xy);
plot(xy, pch=20, cex=3, col="purple4",
   main="Polygon hull around points");
polygon(xyhull, border="purple2", fill="transparent");

}
\seealso{
Other jam plot functions: \code{\link{coordPresets}},
  \code{\link{drawLabels}}, \code{\link{jammaplot}}
}
\concept{jam plot functions}
