% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jam-update-fn-params.R
\name{update_list_elements}
\alias{update_list_elements}
\title{Update a subset of list elements}
\usage{
update_list_elements(
  source_list,
  update_list,
  list_layer_num = 1,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{source_list}{\code{list} with input values}

\item{update_list}{\code{list} with values to populate into \code{source_list}.}

\item{list_layer_num}{\code{integer} used for internal use by this function,
which defines the list depth when traversing a nested list.}

\item{verbose}{\code{logical} indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\description{
Update a subset of list elements
}
\details{
This function is intended to help update a nested \code{source_list},
a subset of whose values should be replaced with entries
in \code{update_list}, leaving any original entries in \code{source_list}
which were not defined in \code{update_list}.

This function may be useful when manipulating lattice or ggplot2
graphical parameters, which are often stored in a nested
list structure.
}
\seealso{
Other jam utilities: 
\code{\link{update_function_params}()}
}
\concept{jam utilities}
